% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{addCyEdge}
\alias{addCyEdge}
\alias{addCyEdgedefunct}
\title{DEFUNCT: addCyEdge}
\usage{
addCyEdgedefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
