% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{addCyEdges}
\alias{addCyEdges}
\title{Add CyEdges}
\usage{
addCyEdges(source.target.list, edgeType = "interacts with",
  directed = FALSE, network = NULL, base.url = .defaultBaseUrl)
}
\description{
Add CyEdges
}
\details{
NOTE: If providing more than one source, target pair, then you
MUST use \code{list} and not \code{c}. See examples.
}
\examples{
\donttest{
addCyEdges(c('sourceNode','targetNode'))
addCyEdges(list('sourceNode','targetNode'))
addCyEdges(list(list('s1','t1'),list('s2','t2')))
}
}
