% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{addCyNode}
\alias{addCyNode}
\alias{addCyNodedefunct}
\title{DEFUNCT: addCyNode}
\usage{
addCyNodedefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
