% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{addGraphToGraph}
\alias{addGraphToGraph}
\alias{addGraphToGraphdefunct}
\title{DEFUNCT: addGraphToGraph}
\usage{
addGraphToGraphdefunct
}
\value{
None
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
