% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{analyzeNetwork}
\alias{analyzeNetwork}
\title{Analyze Network}
\usage{
analyzeNetwork(directed = FALSE, base.url = .defaultBaseUrl)
}
\arguments{
\item{directed}{(optional) If TRUE, the network is considered a directed 
graph. Default is FALSE.}

\item{base.url}{(optional) Ignore unless you need to specify a custom 
domain, port or version to connect to the CyREST API. Default is 
http://localhost:1234 and the latest version of the CyREST API supported 
by this version of RCy3.}
}
\value{
Named list of summary statistics
}
\description{
Calculate various network statistics.
}
\details{
The results are added to the Node and Edge tables and the Results 
Panel. The summary statistics in the Results Panel are also returned by 
the function as a list of named values.
}
\examples{
\donttest{
analyzeNetwork()
analyzeNetwork(TRUE)
}
}
