% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{availableApiVersions}
\alias{availableApiVersions}
\title{Available CyREST API Versions}
\usage{
availableApiVersions(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{list} of API versions as \code{character} strings, e.g., "v1"
}
\description{
Get the list of available CyREST API versions
}
\examples{
\donttest{
availableApiVersions()
# [1] "v1"
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
