% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\docType{data}
\name{CHANNEL}
\alias{CHANNEL}
\title{getBrowserClientChannel}
\format{
An object of class \code{character} of length 1.
}
\usage{
CHANNEL
}
\value{
client channel
}
\description{
Get the unique channel
}
\examples{
\donttest{
getBrowserClientChannel()
}
}
\keyword{datasets}
