% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{checkCytoscapeVersion,OptionalCyObjClass-method}
\alias{checkCytoscapeVersion,OptionalCyObjClass-method}
\title{Check the version of Cytoscape}
\usage{
\S4method{checkCytoscapeVersion}{OptionalCyObjClass}(obj)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
Cytoscape version
}
\description{
Check the version of Cytoscape
}
