% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{checkNotebookIsRunning}
\alias{checkNotebookIsRunning}
\title{checkNotebookIsRunning}
\usage{
checkNotebookIsRunning()
}
\arguments{
\item{newState}{new state of running remote}
}
\description{
checkNotebookIsRunning
}
\examples{
\donttest{
checkNotebookIsRunning()
}
}
