% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{checkRunningRemote}
\alias{checkRunningRemote}
\title{checkRunningRemote}
\usage{
checkRunningRemote(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Determine whether we're running locally or on a remote server. If locally (either via raw R or via a
locally installed Notebook), we prefer to connect to Cytoscape over a local socket. If remote, we have to
connect over Jupyter-Bridge. Either way, we can determine which by whether Cytoscape answers to a version
check. If Cytoscape doesn't answer, we have no information ... and we have to wait until Cytoscape is
started and becomes reachable before we can determine local vs remote.
}
\examples{
\donttest{
checkRunningRemote()
}
}
