% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{clearEdgePropertyBypass}
\alias{clearEdgePropertyBypass}
\title{Clear Edge Property Bypass}
\usage{
clearEdgePropertyBypass(edge.names, visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{List of edge names}

\item{visual.property}{Name of a visual property. See \link{getVisualPropertyNames}.}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Clear bypass values for any edge property of the specified edges,
effectively restoring any previously defined style defaults or mappings.
}
\examples{
\donttest{
clearEdgePropertyBypass()
}
}
\seealso{
\link{setEdgePropertyBypass}
}
