% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDirects.R
\name{clearEdgePropertyDirect}
\alias{clearEdgePropertyDirect}
\title{Clear Edge Property Direct}
\usage{
clearEdgePropertyDirect(edge.names, visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{DESCRIPTION}

\item{visual.property}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
clearEdgePropertyDirect()
}
}
