% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{clearNetworkPropertyBypass}
\alias{clearNetworkPropertyBypass}
\title{Clear Network Property Bypass}
\usage{
clearNetworkPropertyBypass(visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{visual.property}{Name of a visual property. See \link{getVisualPropertyNames}.}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Clear bypass values for any network property,
effectively restoring any previously defined style defaults or mappings.
}
\examples{
\donttest{
clearNetworkPropertyBypass()
}
}
\seealso{
\link{setNodePropertyBypass}
}
