% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{clearNodeOpacityBypass}
\alias{clearNodeOpacityBypass}
\title{Clear Node Opacity Bypass}
\usage{
clearNodeOpacityBypass(node.names, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{List of node names}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Clearn the bypass value for node fill, label and border opacity 
for the specified node or nodes, effectively restoring any previously defined 
style defaults or mappings.
}
\examples{
\donttest{
clearNodeOpacityBypass(c('Node 1','Node 2'))
}
}
\seealso{
{
\link{setNodeOpacityBypass}
}
}
