% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDirects.R
\name{clearNodePropertyDirect}
\alias{clearNodePropertyDirect}
\title{Clear Node Property Direct}
\usage{
clearNodePropertyDirect(node.names, visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{DESCRIPTION}

\item{visual.property}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
clearNodePropertyDirect()
}
}
