% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSelection.R
\name{clearSelection}
\alias{clearSelection}
\title{Clear Selection}
\usage{
clearSelection(type = "both", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{type}{'nodes', 'edges' or 'both' (default)}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
If any nodes are selected in the network, they will be unselected.
}
\examples{
\donttest{
clearSelection()
}
}
\author{
AlexanderPico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
