% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{cloneNetwork,OptionalCyWinClass-method}
\alias{cloneNetwork,OptionalCyWinClass-method}
\title{Copy a Cytoscape Network}
\usage{

  \S4method{cloneNetwork}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork(),
  new.title, return.graph = FALSE)
}
\arguments{
\item{obj}{Cytoscape network}

\item{new.title}{New name for the copy}

\item{return.graph}{Logical whether to copy the graph to a new object in R}
}
\value{
Connection to new copy of network.
}
\description{
Makes a copy of a Cytoscape Network with all of its edges and nodes
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
copy_of_your_net <- cloneNetwork(cw, "new_copy")
}

}
\seealso{
\code{\link{createNetworkFromSelection}}, \code{\link{CytoscapeWindowFromNetwork}}, \code{\link{renameNetwork}}
}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\concept{
RCy3
}
