% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{collapseGroup}
\alias{collapseGroup}
\title{Collapse Group}
\usage{
collapseGroup(groups = NULL, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{groups}{(optional) List of group names or keywords: all, selected,
unselected. Default is the currently selected group.}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Replaces the representation of all of the nodes and edges in a group with a single node.
}
\examples{
\donttest{
collapseGroup()
}
}
