% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandEcho}
\alias{commandEcho}
\title{Command Echo}
\usage{
commandEcho(variable.name = "*", base.url = .defaultBaseUrl)
}
\arguments{
\item{variable.name}{(optional) The name of the variable to display. Default is 
to display all variable values using "*".}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Value of variable
}
\description{
The echo command will display the value of the variable specified 
by the variableName argument, or all variables if variableName is not provided.
}
\examples{
\donttest{
commandEcho()
}
}
