% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandOpenDialog}
\alias{commandOpenDialog}
\title{Command Open Dialog}
\usage{
commandOpenDialog(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
The command line dialog provides a field to enter commands and 
view results. It also provides the help command to display namespaces, 
commands, and arguments
}
\examples{
\donttest{
commandOpenDialog()
}
}
