% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandPause}
\alias{commandPause}
\title{Command Pause}
\usage{
commandPause(message = "", base.url = .defaultBaseUrl)
}
\arguments{
\item{message}{(optional) Text to display in pause dialog}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
The pause command displays a dialog with the text provided in the 
message argument and waits for the user to click OK.
}
\examples{
\donttest{
commandPause('Please click OK to continue.')
}
}
