% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandRun}
\alias{commandRun}
\title{Command Run}
\usage{
commandRun(cmd.string, method = "POST", base.url = .defaultBaseUrl)
}
\arguments{
\item{cmd.string}{(char) command}

\item{method}{HTTP method to be used, e.g., GET or POST (default)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
List object
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function converts a command string into a CyREST URL, executes a GET
request, and parses the HTML result content into an R list object.
}
\examples{
\donttest{
commandRun('layout get preferred')
}
}
