% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandRunFile}
\alias{commandRunFile}
\title{Command Run File}
\usage{
commandRunFile(file, args = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{file}{Path to command script file}

\item{args}{The script arguments as key:value pairs separated by commas}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
The run command will execute a command script from the file pointed 
to by the file argument, which should contain Cytoscape commands, one per line. 
Arguments to the script are provided by the args argument
}
\examples{
\donttest{
commandRunFile('/path/to/my/file.txt')
}
}
