% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterface.R
\name{commandsAPI}
\alias{commandsAPI}
\title{Open Swagger docs for CyREST Commands API}
\usage{
commandsAPI(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
Web page in browser
}
\description{
Opens swagger docs in default browser for a live
instance of Commands available via CyREST.
}
\examples{
\donttest{
commandsAPI()
}
}
