% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandSleep}
\alias{commandSleep}
\title{Command Sleep}
\usage{
commandSleep(duration = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{duration}{(optional) The time in seconds to sleep}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, #' port or version to connect to the CyREST API. Default is http://localhost:1234 #' and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
The sleep command will pause processing for a period of time as 
specified by duration seconds. It is typically used as part of a command script.
}
\examples{
\donttest{
commandSleep(5)
}
}
