% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandsPOST}
\alias{commandsPOST}
\title{Commands POST}
\usage{
commandsPOST(cmd.string, base.url = .defaultBaseUrl)
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
A \code{list}, \code{named list}, \code{status} or None.
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function converts a command string into a CyREST query URL, executes a
POST request, and parses the result content into an R list object.
}
\examples{
\donttest{
commandsPOST('layout get preferred')
commandsPOST('network list properties')
commandsPOST('layout force-directed defaultNodeMass=1')
}
}
