% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandsRun}
\alias{commandsRun}
\title{Run a Command}
\usage{
commandsRun(cmd.string, base.url = .defaultBaseUrl)
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A \code{list}, \code{status} or None.
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function converts a command string into a CyREST query URL, executes a GET
request, and parses the result content into an R list object. Same as commandsGET.
}
\examples{
\donttest{
commandsRun('layout get preferred')
commandsRun('network list properties')
commandsRun('layout force-directed defaultNodeMass=1')
}
}
