% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{connectToNewestCyWindow}
\alias{connectToNewestCyWindow}
\alias{connectToNewestCyWindowdefunct}
\title{DEFUNCT: connectToNewestCyWindow}
\usage{
connectToNewestCyWindowdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
