% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{connectToNewestCyWindow,CytoscapeConnectionClass-method}
\alias{connectToNewestCyWindow,CytoscapeConnectionClass-method}
\title{Creates a connection to the newest Cytoscape window so that it can be further manipulated from R.}
\usage{
\S4method{connectToNewestCyWindow}{CytoscapeConnectionClass}(obj,
  copyToR = FALSE)
}
\arguments{
\item{obj}{Cytoscape network where command is run via RCy3}

\item{copyToR}{If true this copies the graph information to R. This step can be quite slow. Default is false.}
}
\value{
Creates a connection to the newest Cytoscape window so that it can be further manipulated from R
}
\description{
Creates a connection to the newest Cytoscape window so that it can be further manipulated from R.
}
\concept{
RCy3
}

