% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{copyCytoscapeNetwork,CytoscapeWindowClass-method}
\alias{copyCytoscapeNetwork,CytoscapeWindowClass-method}
\title{Copy a Cytoscape Network}
\usage{
\S4method{copyCytoscapeNetwork}{CytoscapeWindowClass}(obj, new_title,
  copy.graph.to.R = FALSE)
}
\arguments{
\item{new_title}{New name for the copy}

\item{copy.graph.to.R}{Logical whether to copy the graph to a new object in R}

\item{object}{Cytoscape network}
}
\value{
Connection to new copy of network.
}
\description{
Makes a copy of a Cytoscape Network with all of its edges and nodes
}
\examples{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
copy_of_your_net <- copyCytoscapeNetwork(cw, "new_copy")
  }
\keyword{graph}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\seealso{
\code{\link{createWindowFromSelection}}, \code{\link{existing.CytoscapeWindow}}, \code{\link{renameCytoscapeNetwork}}
}
\concept{
RCy3
}

