% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createCytoscapejsFromNetwork}
\alias{createCytoscapejsFromNetwork}
\title{Create Cytoscapejs from Network}
\usage{
createCytoscapejsFromNetwork(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(SUID or str or NULL): Name or SUID of a network or view. Default is the
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
(list) The Cytoscape JS object
}
\description{
Create a Cytoscape JS representation of a Cytoscape network
}
\examples{
\donttest{
createCytoscapejsFromNetwork()
}
}
