% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{createGroup}
\alias{createGroup}
\title{Create Group}
\usage{
createGroup(
  group.name,
  nodes = NULL,
  nodes.by.col = "SUID",
  network = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{group.name}{The name used to identify and optionaly label the group}

\item{nodes}{List of node SUIDs, names, other column values, or keyword: 
selected, unselected or all. Default is currently selected nodes.}

\item{nodes.by.col}{name of node table column corresponding to provided nodes 
list. Default is 'SUID'.}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Group SUID
}
\description{
Create a group from the specified nodes.
}
\examples{
\donttest{
createGroup()
}
}
