% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{createGroupByColumn}
\alias{createGroupByColumn}
\title{Create Group by Column}
\usage{
createGroupByColumn(
  group.name,
  column = NULL,
  value = NULL,
  network = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{group.name}{The name used to identify and optionaly label the group}

\item{column}{The name or header of the Node Table column to use for selecting nodes to group}

\item{value}{The value in the column to use for selecting nodes to group}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Group SUID
}
\description{
Create a group of nodes defined by a column value.
}
\examples{
\donttest{
createGroupByColumn('myGroup','Cluster', 'A')
}
}
