% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createNetworkFromGraph}
\alias{createNetworkFromGraph}
\title{Create Network From Graph}
\usage{
createNetworkFromGraph(graph, title = "MyNetwork",
  collection = "myNetworkCollection", base.url = .defaultBaseUrl)
}
\arguments{
\item{graph}{DESCRIPTION}

\item{title}{(char) network name}

\item{collection}{(char) network collection name}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
Creates a Cytoscape network from a Bioconductor graph.
}
\examples{
\donttest{
library(graph)
g <- makeSimpleGraph()
createNetworkFromGraph(g)
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
