% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{createNetworkFromIgraph}
\alias{createNetworkFromIgraph}
\title{Create a Cytoscape network from an igraph network}
\usage{
createNetworkFromIgraph(igraph, new.title = "MyNetwork",
  collection.title = "myNetworkCollection", return.graph = FALSE,
  obj = CytoscapeConnection(), ...)
}
\arguments{
\item{igraph}{(igraph) igraph network object}

\item{new.title}{(char) network name}

\item{collection.title}{(char) network collection name}

\item{...}{params for nodeSet2JSON() and edgeSet2JSON(); see createNetwork}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
(int) network SUID
}
\description{
Takes an igraph network and generates data frames for nodes and edges to
send to the createNetwork function.
Returns the network.suid and applies the perferred layout set in Cytoscape preferences.
}
\details{
Vertices and edges from the igraph network will be translated into nodes and edges
in Cytoscape. Associated attributes will also be passed to Cytoscape as node and edge table columns.
}
\examples{
\donttest{
createNetworkFromIgraph(g)
}
}
\seealso{
createNetworkFromDataFrames, createIgraphFromNetwork
}
