% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{createSubnetwork}
\alias{createSubnetwork}
\title{Create subnetwork from existing network}
\usage{
createSubnetwork(nodes, nodes.by.col = "name", edges, edges.by.col = "name",
  exclude.edges = "F", new.title, return.graph = FALSE,
  obj = CytoscapeWindowFromNetwork())
}
\arguments{
\item{nodes}{list of node names or keyword: selected, unselected or all}

\item{nodes.by.col}{name of node table column corresponding to provided nodes list; default is 'name'}

\item{edges}{list of edge names or keyword: selected, unselected or all}

\item{edges.by.col}{name of edge table column corresponding to provided edges list; default is 'name'}

\item{exclude.edges}{(boolean) whether to exclude connecting edges; default is FALSE}

\item{new.title}{name of new subnetwork to be created;
default is to add a numbered suffix to source network name}

\item{network}{name or suid of the source network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
SUID of new subnetwork
}
\description{
Copies a subset of nodes and edges into a newly created subnetwork.
}
\details{
If you specify both nodes and edges, the resulting subset will be the union of those sets.
Typical usage only requires specifying either nodes or edges. Note that selected nodes will bring
along their connecting edges by default (see exclude.edges arg) and selected edges will always
bring along their source and target nodes.
}
\examples{
\donttest{
createSubnetwork("selected")
createSubnetwork("selected",new.title="mySubnetwork")
createSubnetwork(c("node 1","node 2","node 3"))
createSubnetwork(c("AKT1","TP53","PIK3CA"),"display name")
createSubnetwork(edges="all") #subnetwork of all connected nodes
}
}
\seealso{
createNetworkFromDataFrames
}
