% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{createWindow}
\alias{createWindow}
\title{DEPRECATED: createWindow}
\usage{
Use the replacement function instead:
\link{createNetworkFromGraph}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
