% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{createWindowFromSelection}
\alias{createWindowFromSelection}
\title{DEPRECATED: createWindowFromSelection}
\usage{
Use the replacement function instead:
\code{\link[RCy3]{createNetworkFromSelection}}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
