% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{cy2.edge.names}
\alias{cy2.edge.names}
\alias{cy2.edge.namesdefunct}
\title{DEFUNCT: cy2.edge.names}
\usage{
cy2.edge.namesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
