% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserClose}
\alias{cybrowserClose}
\title{Cybrowser Close}
\usage{
cybrowserClose(id = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{id}{(optional) The identifier for the browser window to close}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Close an internal web browser and remove all content. Provide an 
id for the browser you want to close.
}
\examples{
\donttest{
cybrowserClose('554')
}
}
