% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserHide}
\alias{cybrowserHide}
\title{Cybrowser Hide}
\usage{
cybrowserHide(id = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{id}{(optional) The identifier for the browser window to hide}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Hide an existing browser, whether it's in the Results panel or a separate window.
}
\examples{
\donttest{
cybrowserHide()
}
}
\seealso{
\link{cybrowserShow}

\link{cybrowserDialog}
}
