% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserList}
\alias{cybrowserList}
\title{Cybrowser List}
\usage{
cybrowserList(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of open cybrowser windows
}
\description{
List all browsers that are currently open, whether as a dialog or in the results panel.
}
\examples{
\donttest{
cybrowserList()
}
}
