% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserSend}
\alias{cybrowserSend}
\title{Cybrowser Send}
\usage{
cybrowserSend(id = NULL, script, base.url = .defaultBaseUrl)
}
\arguments{
\item{id}{(optional) The identifier for the browser window}

\item{script}{(optional) A string that represents a JavaScript variable, script,
or call to be executed in the browser. Note that only string results are returned.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
String result
}
\description{
Send the text to the browser indicated by the id and return the 
response, if any. Note that the JSON result field could either be a bare string 
or JSON formatted text.
}
\examples{
\donttest{
cybrowserSend(id="Window 1", script="navigator.userAgent;")
}
}
