% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserShow}
\alias{cybrowserShow}
\title{Cybrowser Show}
\usage{
cybrowserShow(id = NULL, text = NULL, title = NULL, url = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{id}{(optional) The identifier for the browser window}

\item{text}{(optional) HTML text to initially load into the browser}

\item{title}{(optional) Text to be shown in the title bar of the browser window}

\item{url}{(optional) The URL the browser should load}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Launch Cytoscape's internal web browser in a pane in the Result 
Panel. Provide an id for the window if you want subsequent control of the 
window via cybrowser hide.
}
\examples{
\donttest{
cybrowserShow(url='http://cytoscape.org')
}
}
\seealso{
\link{cybrowserDialog}

\link{cybrowserHide}
}
