% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tools.R
\name{cybrowserVersion}
\alias{cybrowserVersion}
\title{Cybrowser Version}
\usage{
cybrowserVersion(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain, 
port or version to connect to the CyREST API. Default is http://localhost:1234 
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Version number
}
\description{
Display the version of the CyBrowser app.
}
\examples{
\donttest{
cybrowserVersion()
}
}
