% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestDELETE}
\alias{cyrestDELETE}
\title{CyREST DELETE}
\usage{
cyrestDELETE(operation = NULL, parameters = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{operation}{A string to be converted to the REST query namespace}

\item{parameters}{A named list of values to be converted to REST query parameters}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
CyREST result content
}
\description{
Constructs the query, makes DELETE call and processes the result
}
\examples{
\donttest{
cyrestDELETE('session')
}
}
