% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestGET}
\alias{cyrestGET}
\title{CyREST GET}
\usage{
cyrestGET(operation = NULL, parameters = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{operation}{DESCRIPTION}

\item{parameters}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
cyrestGET()
}
}
