% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestPOST}
\alias{cyrestPOST}
\title{CyREST POST}
\usage{
cyrestPOST(operation, parameters = NULL, body = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{operation}{DESCRIPTION}

\item{parameters}{DESCRIPTION}

\item{body}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
cyrestPOST()
}
}
