% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestPUT}
\alias{cyrestPUT}
\title{CyREST PUT}
\usage{
cyrestPUT(operation, parameters = NULL, body = FALSE,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{operation}{DESCRIPTION}

\item{parameters}{DESCRIPTION}

\item{body}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
cyrestPUT()
}
}
