% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeFreeMemory}
\alias{cytoscapeFreeMemory}
\title{Free Up Unused Memory for Cytoscape}
\usage{
cytoscapeFreeMemory(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
status message
}
\description{
Manually call Java's garbage collection \code{System.gc()} to free up unused memory. 
This process happens automatically, but may be useful to call explicitly for testing or evaluation purposes.
}
\examples{
\donttest{
cytoscapeFreeMemory()
# [1] "Unused memory freed up."
}
}
\author{
Alexander Pico
}
