% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeMemoryStatus}
\alias{cytoscapeMemoryStatus}
\title{Memory Available to Cytoscape}
\usage{
cytoscapeMemoryStatus(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{list} of \code{numeric} values
}
\description{
Returns the memory resources of the server running Cytoscape
}
\examples{
\donttest{
cytoscapeMemoryStatus()
#  usedMemory  freeMemory totalMemory   maxMemory 
#         181        2624        2805       13653 
}
}
\author{
Alexander Pico
}
