% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeNumberOfCores}
\alias{cytoscapeNumberOfCores}
\title{Number of Processors Available to Cytoscape}
\usage{
cytoscapeNumberOfCores(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{numeric} value
}
\description{
Returns the processor resources of the server running Cytoscape
}
\examples{
\donttest{
cytoscapeNumberOfCores()
# [1] 8
}
}
\author{
Alexander Pico
}
