% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeVersionInfo}
\alias{cytoscapeVersionInfo}
\title{Cytoscape and CyREST API Versions}
\usage{
cytoscapeVersionInfo(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
\code{list} of versions
}
\description{
Returns the versions of the current Cytoscape and CyREST API
}
\examples{
\donttest{
cytoscapeVersionInfo()
 # apiVersion cytoscapeVersion 
 #       "v1" "3.7.0-SNAPSHOT" 
}
}
\author{
Alexander Pico
}
