% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{CytoscapeWindow}
\alias{CytoscapeWindow}
\title{DEPRECATED: CytoscapeWindow}
\usage{
Use the replacement function instead:
\code{createNetworkFromGraph}
Rely on the default value or overwrite with custom host and port details.
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
