% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{deleteAnnotation}
\alias{deleteAnnotation}
\title{Delete Annotation}
\usage{
deleteAnnotation(names = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{names}{List of annotations by UUID or Name}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is 
http://localhost:1234 and the latest version of the CyREST API supported by 
this version of RCy3.}
}
\value{
None
}
\description{
Remove an annotation from the current network view in Cytoscape
}
\examples{
\donttest{
deleteAnnotation("016a4af1-69bc-4b99-8183-d6f118847f96")
deleteAnnotation(c("T2","T3"))
deleteAnnotation(sapply(getAnnotationList(), '[[', 'uuid'))
}
}
