% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSelection.R
\name{deleteDuplicateEdges}
\alias{deleteDuplicateEdges}
\title{Delete Duplicate Edges}
\usage{
deleteDuplicateEdges(
  network = NULL,
  base.url = .defaultBaseUrl,
  ignoreDirection = FALSE
)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}

\item{ignoreDirection}{(optional)  True to treat x->y as equal to y->x.
Default is FALSE.}
}
\value{
Lists of SUIDs for selected nodes and edges
}
\description{
Removes edges with duplicate names. Only considers cases with
identical source, target, interaction and directionality.
}
\details{
Duplicate edges are first selected and then deleted. Prior edge
selections will be lost; node selections will not be affected.
}
\examples{
\donttest{
deleteDuplicateEdges()
}
}
