% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{deleteEdgeAttribute}
\alias{deleteEdgeAttribute}
\alias{deleteEdgeAttribute_deprecated}
\title{DEPRECATED: deleteEdgeAttribute}
\usage{
deleteEdgeAttribute_deprecated
}
\value{
None
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
Use the replacement function instead: 
\link{deleteTableColumn}
}
