% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{deleteGroup}
\alias{deleteGroup}
\title{Delete (or Ungroup) a Group}
\usage{
deleteGroup(groups = NULL, groups.by.col = "SUID", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{groups}{(optional) List of group SUIDs, names, other column values or keywords: all, selected,
unselected. Default is the currently selected group.}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}

\item{nodes.by.col}{name of node table column corresponding to provided groups 
list. Default is 'SUID'.}
}
\value{
None
}
\description{
Deletes one or more groups, while leaving member nodes intact.
}
\details{
Note: group nodes are ungrouped but not deleted in Cytoscape version 3.6.1
}
\examples{
\donttest{
deleteGroup()
}
}
