% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{deleteNetwork}
\alias{deleteNetwork}
\title{Delete Network}
\usage{
deleteNetwork(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Delete a network from the current Cytoscape session.
}
\examples{
\donttest{
deleteNetwork()
}
}
