% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{deleteSelectedNodes}
\alias{deleteSelectedNodes}
\title{Delete Selected Nodes}
\usage{
deleteSelectedNodes(network = NULL, base.url = .defaultBaseUrl)
}
\value{
A \code{named list} of deleted node suids ($nodes) as well as edge suids 
($edges) deleted as a result of the node deletion
}
\description{
Delete Selected Nodes
}
