% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{deleteTableColumn}
\alias{deleteTableColumn}
\title{Delete a table column}
\usage{
deleteTableColumn(column, table = "node", namespace = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{Name of table, e.g., node (default), edge, network}

\item{namespace}{Namespace of table. Default is "default".}

\item{network}{Name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}

\item{columns}{Name of the column to delete}
}
\value{
A \code{data.frame} of column values
}
\description{
Delete a column from node, edge or network tables.
}
\examples{
\donttest{
deleteTableColumn('node','group')
}
}
