% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{deleteVisualStyle}
\alias{deleteVisualStyle}
\title{Delete Visual Style}
\usage{
deleteVisualStyle(style.name, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{(char) name of style to delete}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Deletes the specified visual style from current session.
}
\examples{
\donttest{
deleteVisualStyle("myStyle")
}
}
