% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{demoSimpleGraph}
\alias{demoSimpleGraph}
\alias{demoSimpleGraphdefunct}
\title{DEFUNCT: demoSimpleGraph}
\usage{
demoSimpleGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
