% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{displayGraph}
\alias{displayGraph}
\alias{displayGraphdefunct}
\title{DEFUNCT: displayGraph}
\usage{
displayGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
