% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{doInitializeSandbox}
\alias{doInitializeSandbox}
\title{doInitializeSandbox}
\usage{
doInitializeSandbox(requester = NULL, base.url = .defaultBaseUrl)
}
\description{
doInitializeSandbox
}
\examples{
\donttest{
doInitializeSandbox()
}
}
