% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{doRequestRemoteCommands}
\alias{doRequestRemoteCommands}
\title{doRequestRemoteCommands}
\usage{
doRequestRemoteCommands(method, qurl, qbody = NULL, headers = NULL)
}
\description{
Do requests remotely
}
\examples{
\donttest{
doRequestRemoteCommands()
}
}
