% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{doRequestWrapper}
\alias{doRequestWrapper}
\title{doRequestWrapper}
\usage{
doRequestWrapper(method, qurl, qbody = NULL, ...)
}
\description{
doRequestWrapper
}
\examples{
\donttest{
doRequestWrapper()
}
}
