% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{eda.names}
\alias{eda.names}
\alias{eda.namesdefunct}
\title{DEFUNCT: eda.names}
\usage{
eda.namesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
