% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated.R
\name{existing.CytoscapeWindow}
\alias{existing.CytoscapeWindow}
\title{Deprecated: existing.CytoscapeWindow}
\usage{
Use the replacement function instead:
\code{\link[RCy3]{CytoscapeWindowFromNetwork}}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
