% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{exportFilters}
\alias{exportFilters}
\title{Export Filters}
\usage{
exportFilters(filename = "filters.json", base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) Full path or path relavtive to current 
working directory, in addition to the name of the file. Default is 
"filters.json"}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Saves filters to file in JSON format.
}
\details{
Unlike other export functions, Cytoscape will automatically
overwrite files with the same name. You will not be prompted to confirm
or reject overwrite. Use carefully!
}
\examples{
\donttest{
exportFilters()
}
}
