% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{exportImage}
\alias{exportImage}
\title{Exports the current network view as an image}
\usage{
exportImage(file.name, type, resolution = 300, units = "inches",
  height = 600, width = 800, zoom = 200, base.url = .defaultBaseUrl)
}
\arguments{
\item{type}{(char) type of image to export, e.g., JPEG, PDF, PNG, PostScript, SVG (case sensitive)}

\item{base.url}{cyrest base url for communicating with cytoscape}

\item{filename}{(char) name of the image file to save}
}
\value{
server response
}
\description{
Exports the current network view as an image
}
\details{
The image is cropped per the current view in Cytoscape.
}
\examples{
\donttest{
exportImage('myNetwork','PDF')
}
}
