% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{exportVisualStyles}
\alias{exportVisualStyles}
\title{Export Visual Styles}
\usage{
exportVisualStyles(filename = NULL, type = NULL, styles = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) Name of the style file to save. Default is "styles.xml"}

\item{type}{(optional) Type of data file to export, e.g., XML, JSON (case sensitive). 
Default is XML.}

\item{styles}{(optional) The styles to be exported, listed as a comma-separated string. 
If no styles are specified, only the current one is exported.}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
None
}
\description{
Saves one or more visual styles to file
}
\examples{
\donttest{
exportVisualStyles('/fullpath/myStyle')
exportVisualStyles('/fullpath/myStyle', type = 'JSON')
exportVisualStyles('/fullpath/myStyle', style = 'Minimal,default,Directed')
}
}
