% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{findRemoteCytoscape}
\alias{findRemoteCytoscape}
\title{findRemoteCytoscape}
\usage{
findRemoteCytoscape(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Figure out whether CyREST is local or remote. If remote, we'll 
want to go through Jupyter-Bridge.
}
\examples{
\donttest{
findRemoteCytoscape()
}
}
