% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{fitContent}
\alias{fitContent}
\title{Fit Content}
\usage{
fitContent(selected.only = FALSE, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{selected.only}{(Boolean) Whether to fit only current selection. Default 
is false, i.e., to fit the entire network.}

\item{network}{(optional) Name or SUID of a network or view. Default is the 
"current" network active in Cytoscape. The first (presummably only) view 
associated a network is used if a specific view SUID is not provided.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Zoom and pan network view to maximize either height or width of
current network window.
}
\details{
Takes first (presumably only) view associated with provided network
}
\examples{
\donttest{
fitContent()
}
}
