% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterface.R
\name{floatPanel}
\alias{floatPanel}
\title{Float Panel}
\usage{
floatPanel(panel.name, base.url = .defaultBaseUrl)
}
\arguments{
\item{panel.name}{Name of the panel. Multiple ways of referencing panels is supported:\cr
WEST, control panel, control, c \cr
SOUTH, table panel, table, ta \cr
SOUTH_WEST, tool panel, tool, to \cr
EAST, results panel, results, r}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Pop out a panel from the UI of Cytoscape. Other panels will 
expand into the space.
}
\examples{
\donttest{
floatPanel('table')
}
}
