% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getAdjacentEdgeNames}
\alias{getAdjacentEdgeNames}
\alias{getAdjacentEdgeNamesdefunct}
\title{DEFUNCT: getAdjacentEdgeNames}
\usage{
getAdjacentEdgeNamesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
Use these replacement functions instead: 
\link{selectEdgesAdjacentToSelectedNodes}, \link{getSelectedEdges}
}
