% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getAllEdgeAttributes}
\alias{getAllEdgeAttributes}
\alias{getAllEdgeAttributes_deprecated}
\title{DEFUNCT: getAllEdgeAttributes}
\usage{
getAllEdgeAttributes_deprecated
}
\value{
None
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
Use the replacement function instead: 
\link{getTableColumns}
}
