% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{getAllStyleMappings}
\alias{getAllStyleMappings}
\title{Get All Style Mappings}
\usage{
getAllStyleMappings(style.name = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{Name for style; default is the 'default' style}

\item{base_url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is 
http://localhost:1234 and the latest version of the CyREST API supported by 
this version of RCy3.}
}
\value{
List of lists with named values defining the visual property mappings
}
\description{
Retrieve the values the define all the mappings per visual 
property in a given style.
}
\examples{
\donttest{
getStyleMapping()
}
}
