% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotations.R
\name{getAnnotationList}
\alias{getAnnotationList}
\title{Get Annotation List}
\usage{
getAnnotationList(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is 
http://localhost:1234 and the latest version of the CyREST API supported by 
this version of RCy3.}
}
\value{
RETURN_DESCRIPTION
}
\description{
A list of named lists with annotation information
}
\details{
You can obtain a list of UUIDs by applying a subset function
like so: sapply(getAnnotationList(), '[[', 'uuid')
}
\examples{
\donttest{
getAnnotationList()
}
}
