% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getAppStatus}
\alias{getAppStatus}
\title{App Status}
\usage{
getAppStatus(app, base.url = .defaultBaseUrl)
}
\arguments{
\item{app}{Name of app}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
App name and status
}
\description{
Retrieve the current status of a Cytoscape app: Installed, Uninstalled
or Disabled.
}
\examples{
\donttest{
getAppStatus()
}
}
