% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getAppUpdates}
\alias{getAppUpdates}
\title{List Apps With Updates}
\usage{
getAppUpdates(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A list of updatable app names, versions and statuses
}
\description{
Retrieve list of currently installed Cytoscape apps with updates available.
}
\examples{
\donttest{
getAppUpdates()
}
}
