% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getArrowShapes}
\alias{getArrowShapes}
\title{Get Arrow Shapes}
\usage{
getArrowShapes(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
A list of \code{character} strings, e.g., 'DIAMOND', 'T', 'ARROW'
}
\description{
Retrieve the names of the currently supported 'arrows' -- the
decorations can (optionally) appear at the ends of edges, adjacent to
the nodes they connect, and conveying information about the nature of
the nodes' relationship.
}
\examples{
\donttest{
getArrowShapes()
}
}
