% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getAttributeClassNames}
\alias{getAttributeClassNames}
\alias{getAttributeClassNamesdefunct}
\title{DEFUNCT: getAttributeClassNames}
\usage{
getAttributeClassNamesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
