% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getAvailableApps}
\alias{getAvailableApps}
\title{List Available Apps}
\usage{
getAvailableApps(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getAvailableApps()
}
}
