% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{getBrowserClientChannel}
\alias{getBrowserClientChannel}
\title{getBrowserClientChannel}
\usage{
getBrowserClientChannel()
}
\value{
client channel
}
\description{
Get the unique channel
}
\examples{
\donttest{
getBrowserClientChannel()
}
}
