% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{getBrowserClientJs}
\alias{getBrowserClientJs}
\title{getBrowserClientJs}
\usage{
getBrowserClientJs()
}
\value{
Javascript inject code
}
\description{
Prepend channel number of client Javascript so it can communicate with this process via Jupyter-bridge
}
\examples{
\donttest{
getBrowserClientJs()
}
}
