% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collections.R
\name{getCollectionList}
\alias{getCollectionList}
\title{Get Collection List}
\usage{
getCollectionList(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getCollectionList()
}
}
