% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collections.R
\name{getCollectionNetworks}
\alias{getCollectionNetworks}
\title{Get Collection Networks}
\usage{
getCollectionNetworks(collection.suid = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{collection.suid}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getCollectionNetworks()
}
}
