% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collections.R
\name{getCollectionSuid}
\alias{getCollectionSuid}
\title{Get Collection Suid}
\usage{
getCollectionSuid(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getCollectionSuid()
}
}
