% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{getCommandNames,CytoscapeConnectionClass-method}
\alias{getCommandNames,CytoscapeConnectionClass-method}
\title{Gets commands available from within cytoscape from 
functions within cytoscape and from installed plugins.}
\usage{
\S4method{getCommandNames}{CytoscapeConnectionClass}(obj)
}
\arguments{
\item{obj}{Cytoscape network where commands are fetched via RCy3}
}
\value{
Vector of available commands from all namespaces (e.g. functions and plugins)
}
\description{
Gets commands available from within cytoscape from 
functions within cytoscape and from installed plugins.
}
\concept{
RCy3
}

