% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getCommandNamesWithinNamespace}
\alias{getCommandNamesWithinNamespace}
\alias{getCommandNamesWithinNamespace_deprecated}
\title{DEFUNCT: getCommandNamesWithinNamespace}
\usage{
getCommandNamesWithinNamespace_deprecated
}
\value{
None
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
Use the replacement function instead: 
\link{commandsHelp}
}
