% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{getCommandsWithinNamespace,CytoscapeConnectionClass-method}
\alias{getCommandsWithinNamespace,CytoscapeConnectionClass-method}
\title{Gets commands available from within a namespace in Cytoscape from 
functions within cytoscape and from installed plugins.}
\usage{
\S4method{getCommandsWithinNamespace}{CytoscapeConnectionClass}(obj, namespace)
}
\arguments{
\item{obj}{Cytoscape network where commands are fetched via RCy3}

\item{namespace}{Cytoscape function (e.g. layout or network settings) or Cytoscape plugin function}
}
\value{
Vector of available commands from a specific plugin or Cytoscape function (e.g. namespace)
}
\description{
Gets commands available from within a namespace in Cytoscape from 
functions within cytoscape and from installed plugins.
}
\concept{
RCy3
}

