% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getCurrentSandbox}
\alias{getCurrentSandbox}
\title{getCurrentSandbox}
\usage{
getCurrentSandbox()
}
\value{
current sandbox
}
\description{
getCurrentSandbox
}
\examples{
\donttest{
getCurrentSandbox()
}
}
