% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getCurrentSandboxName}
\alias{getCurrentSandboxName}
\title{getCurrentSandboxName}
\usage{
getCurrentSandboxName()
}
\value{
current sandbox name
}
\description{
getCurrentSandboxName
}
\examples{
\donttest{
getCurrentSandboxName()
}
}
