% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getCurrentSandboxPath}
\alias{getCurrentSandboxPath}
\title{getCurrentSandboxPath}
\usage{
getCurrentSandboxPath()
}
\value{
current sandbox path
}
\description{
Return the current sandbox path.
}
\examples{
\donttest{
getCurrentSandboxPath()
}
}
