% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getDefaultEdgeReverseSelectionColor}
\alias{getDefaultEdgeReverseSelectionColor}
\alias{getDefaultEdgeReverseSelectionColordefunct}
\title{DEFUNCT: getDefaultEdgeReverseSelectionColor}
\usage{
getDefaultEdgeReverseSelectionColordefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
