% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getDefaultNodeReverseSelectionColor}
\alias{getDefaultNodeReverseSelectionColor}
\alias{getDefaultNodeReverseSelectionColordefunct}
\title{DEFUNCT: getDefaultNodeReverseSelectionColor}
\usage{
getDefaultNodeReverseSelectionColordefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
