% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getDefaultSandboxPath}
\alias{getDefaultSandboxPath}
\title{getDefaultSandboxPath}
\usage{
getDefaultSandboxPath()
}
\value{
default sandbox path
}
\description{
getDefaultSandboxPath
}
\examples{
\donttest{
getDefaultSandboxPath()
}
}
