% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getDisabledApps}
\alias{getDisabledApps}
\title{List Disabled Apps}
\usage{
getDisabledApps(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getDisabledApps()
}
}
