% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleValues.R
\name{getEdgeColor}
\alias{getEdgeColor}
\title{Get Edge Color}
\usage{
getEdgeColor(edge.names = NULL, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getEdgeColor()
}
}
