% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{getEdgeSelectionColorDefault}
\alias{getEdgeSelectionColorDefault}
\title{Get Edge Selection Color Default}
\usage{
getEdgeSelectionColorDefault(style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Retrieve the default selected edge color.
}
\examples{
\donttest{
getEdgeSelectionColorDefault()
}
}
