% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{getFilterList}
\alias{getFilterList}
\title{Get Filter List}
\usage{
getFilterList(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of filter names
}
\description{
Retrieve list of named filters in current session.
}
\examples{
\donttest{
getFilterList()
}
}
