% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getFirstNeighbors}
\alias{getFirstNeighbors}
\title{Get list of nodes neighboring provided list}
\usage{
getFirstNeighbors(node.names, as.nested.list = FALSE, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{A \code{list} of node names from the \code{name} column of the \code{node table}}

\item{as.nested.list}{\code{logical} Whether to return lists of neighbors per query node}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
A list of unique node names, optionally nested per query node name.
}
\description{
Returns a non-redundan list of first
neighbors of the supplied list of nodes.
}
\examples{
/dontrun {
}
}
\seealso{
selectNodes  
selectFirstNeighbors
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
