% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getGraph}
\alias{getGraph}
\alias{getGraphdefunct}
\title{DEFUNCT: getGraph}
\usage{
getGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{createGraphFromNetwork}
}
