% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{getGraphFromNetwork,missing-method}
\alias{getGraphFromNetwork,missing-method}
\alias{getGraphFromNetwork,CytoscapeConnectionClass-method}
\alias{getGraphFromNetwork,CytoscapeWindowClass-method}
\title{getGraphFromNetwork}
\usage{
\S4method{getGraphFromNetwork}{missing}(title = NA)

\S4method{getGraphFromNetwork}{CytoscapeConnectionClass}(obj, title = NA)

\S4method{getGraphFromNetwork}{CytoscapeWindowClass}(obj)
}
\value{
A Bioconductor graph object.
}
\description{
Returns the Cytoscape network as a Bioconductor graph.
}
\examples{
\donttest{cw <- CytoscapeWindow('network', graph=makeSimpleGraph())
displayGraph(cw)
layoutNetwork(cw)
g.net1 <- getGraphFromNetwork(cw)

cc <- CytoscapeConnection()
g.net2 <- getGraphFromNetwork(cc, 'network')

g.net3 <- getGraphFromNetwork('network') #default connection

g.net4 <- getGraphFromNetwork() #current network
}
}
\author{
Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
