% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{getGroupInfo}
\alias{getGroupInfo}
\title{Get Group Information}
\usage{
getGroupInfo(group, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{group}{(optional) Group name or SUID.}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Group information
}
\description{
Retrieve information about a group by providing a network and the group node identifier.
}
\examples{
\donttest{
getGroupInfo('mygroup')
}
}
