% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getInstalledApps}
\alias{getInstalledApps}
\title{List Installed Apps}
\usage{
getInstalledApps(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getInstalledApps()
}
}
