% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{getJupyterBridgeURL}
\alias{getJupyterBridgeURL}
\title{getJupyterBridgeURL}
\usage{
getJupyterBridgeURL()
}
\value{
jupyter bridge server url
}
\description{
Get the jupyter bridge server url
}
\examples{
\donttest{
getJupyterBridgeURL()
}
}
