% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Layouts.R
\docType{methods}
\name{getLayoutNames,missing-method}
\alias{getLayoutNames,missing-method}
\alias{getLayoutNames,character-method}
\title{Get Layout Names}
\usage{
\S4method{getLayoutNames}{missing}()

\S4method{getLayoutNames}{character}(base.url)
}
\value{
A \code{list} of \code{character} strings, e.g., "force-directed" "circular" "grid"
}
\description{
Retrieve the names of the currently supported layout algorithms.  These
may be used in subsequent calls to the 'layoutNetwork' function.  Note that
some of the more attractive layout options, from yFiles, cannot be
run except from the user interface; their names do not appear here.
}
