% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetworkCount}
\alias{getNetworkCount}
\title{Get the number of Cytoscape networks}
\usage{
getNetworkCount(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
\code{numeric}
}
\description{
Returns the number of Cytoscape networks in the current Cytoscape session
}
\examples{
\dontrun{
getNetworkCount()
# 3
}
}
\author{
Alexander Pico Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
