% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetworkList}
\alias{getNetworkList}
\title{Get the list of Cytoscape networks}
\usage{
getNetworkList(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
\code{list}
}
\description{
Returns the list of Cytoscape network names in the current Cytoscape session
}
\examples{
\dontrun{
getNetworkList()
# 3
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
