% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{getNetworkName,OptionalCyObjClass-method}
\alias{getNetworkName,OptionalCyObjClass-method}
\title{Get the name of a network}
\usage{
\S4method{getNetworkName}{OptionalCyObjClass}(network.suid = NA,
  obj = CytoscapeConnection())
}
\arguments{
\item{network.suid}{SUID of the network; default is current network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network name
}
\description{
Get the name of a network
}
\examples{
\donttest{
getNetworkName()
getNetworkName(1111)
}
}
