% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{getNetworkName}
\alias{getNetworkName}
\title{Get the name of a network}
\usage{
getNetworkName(network.suid = NA, obj = CytoscapeConnection())
}
\arguments{
\item{network.suid}{SUID of the network; default is current network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network viewid
}
\description{
Get the name of a network
}
\examples{
\donttest{
getNetworkName()
getNetworkName(1111)
}
}
