% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CyNDEx.R
\name{getNetworkNDExId}
\alias{getNetworkNDExId}
\title{Get Network NDEx Id}
\usage{
getNetworkNDExId(network = NULL, base.url = .cyndexBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the "current" 
network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
NDEx identifier (\code{externalId}) or NULL
}
\description{
Retrieve the NDEx externalId for a Cytoscape network, presuming
it has already been exported to NDEx.
}
\details{
If the Cytoscape network is not associated with an NDEx network, the
return value will be NULL.
}
\examples{
\donttest{
getNetworkNDExId()
}
}
