% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{getNetworkViews}
\alias{getNetworkViews}
\title{Get Network Views}
\usage{
getNetworkViews(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getNetworkViews()
}
}
