% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{getNetworkViewSuid}
\alias{getNetworkViewSuid}
\title{Get the SUID of a network view}
\usage{
getNetworkViewSuid(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{Name or SUID of the network; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
(\code{numeric}) Network view suid
}
\description{
Retrieve the SUID of a network view
}
\examples{
\donttest{
getNetworkViewSuid()
getNetworkViewSuid("myNetwork")
# 90
}
}
\author{
Alexander Pico
}
