% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleValues.R
\name{getNetworkZoom}
\alias{getNetworkZoom}
\title{Get Network Zoom}
\usage{
getNetworkZoom(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getNetworkZoom()
}
}
