% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleValues.R
\name{getNodeColor}
\alias{getNodeColor}
\title{Get Node Color}
\usage{
getNodeColor(node.names = NULL, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getNodeColor()
}
}
