% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getNodeShapes}
\alias{getNodeShapes}
\title{Get Node Shapes}
\usage{
getNodeShapes(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A list of \code{character} strings, e.g. 'ELLIPSE', 'RECTANGLE'
}
\description{
Retrieve the names of the currently supported node shapes,
which can then be used in calls to setNodeShapeRule and
setDefaultVizMapValue.
}
\examples{
\donttest{
getNodeShapes()
}
}
