% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleValues.R
\name{getNodeSize}
\alias{getNodeSize}
\title{Get Node Size}
\usage{
getNodeSize(node.names = NULL, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{List of node names}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Property value
}
\description{
Retrieve the actual size of specified nodes.
}
\examples{
\donttest{
getNodeSize()
}
}
