% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{getNotebookIsRunning}
\alias{getNotebookIsRunning}
\title{getNotebookIsRunning}
\usage{
getNotebookIsRunning()
}
\value{
None
}
\description{
getNotebookIsRunning
}
\examples{
\donttest{
getNotebookIsRunning()
}
}
