% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getSandboxReinitialize}
\alias{getSandboxReinitialize}
\title{getSandboxReinitialize}
\usage{
getSandboxReinitialize()
}
\value{
sandbox reinitialize
}
\description{
getSandboxReinitialize
}
\examples{
\donttest{
getSandboxReinitialize()
}
}
