% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSelection.R
\name{getSelectedNodes}
\alias{getSelectedNodes}
\title{Get Selected Nodes}
\usage{
getSelectedNodes(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
\code{list} of selected node names
}
\description{
Retrieve the names of all the nodes selected in the network.
}
\examples{
\donttest{
getSelectedNodes()
}
}
\author{
AlexanderPico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
