% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{getStyleDependencies}
\alias{getStyleDependencies}
\title{Get the values of dependencies in a style}
\usage{
getStyleDependencies(style.name, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{(char) name for style}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Updates style dependencies, overriding any prior settings.
}
\section{Available Dependencies}{

arrowColorMatchesEdge
nodeCustomGraphicsSizeSync
nodeSizeLocked
}

\examples{
\donttest{
getStyleDependencies("myStyle")
}
}
