% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{getStyleMapping}
\alias{getStyleMapping}
\title{Get Style Mapping}
\usage{
getStyleMapping(visual.prop, style.name = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{visual.prop}{Name of the visual property, e.g., NODE_FILL_COLOR}

\item{style.name}{Name for style; default is the 'default' style}

\item{base_url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is 
http://localhost:1234 and the latest version of the CyREST API supported by 
this version of RCy3.}
}
\value{
List of named values defining the visual property mappings
}
\description{
Retrieve the values the define the mappings for a given
visual property in a style.
}
\examples{
\donttest{
getStyleMapping()
}
}
\seealso{
getVisualPropertyNames
}
