% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{getTableColumnNames}
\alias{getTableColumnNames}
\title{Get Table Column Names}
\usage{
getTableColumnNames(table = "node", namespace = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{name of table, e.g., node, edge, network; default is "node"}

\item{namespace}{namespace of table, e.g., default}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of column names
}
\description{
Retrieve the names of all columns in a table
}
\examples{
\donttest{
getTableColumnNames()
getTableColumnNames('edge')
getTableColumnNames('network')
}
}
