% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{getTableColumnTypes}
\alias{getTableColumnTypes}
\title{Get Table Column Types}
\usage{
getTableColumnTypes(table = "node", namespace = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{name of table, e.g., node, edge, network; default is "node"}

\item{namespace}{namespace of table, e.g., default}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
a named list of column types
}
\description{
Retrieve the types of all columns in a table
}
\examples{
\donttest{
getTableColumnTypes()
getTableColumnTypes('edge')
getTableColumnTypes('network')
}
}
