% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getUninstalledApps}
\alias{getUninstalledApps}
\title{List Uninstalled Apps}
\usage{
getUninstalledApps(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getUninstalledApps()
}
}
