% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{getVersionInfo,missing-method}
\alias{getVersionInfo,missing-method}
\alias{getVersionInfo,CytoscapeConnectionClass-method}
\title{Get Cytoscape and CyREST API Versions}
\usage{
\S4method{getVersionInfo}{missing}()

\S4method{getVersionInfo}{CytoscapeConnectionClass}(obj)
}
\arguments{
\item{obj}{(optional) \code{CytoscapeConnection}}
}
\value{
\code{list} of versions
}
\description{
Get the running versions of Cytoscape and the CyREST API
}
\examples{
\donttest{
getVersionInfo()

cy <- CytoscapeConnection()
getVersionInfo(cy)
}
}
\author{
Alexander Pico
}
