% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getViewCoordinates}
\alias{getViewCoordinates}
\alias{getViewCoordinatesdefunct}
\title{DEFUNCT: getViewCoordinates}
\usage{
getViewCoordinatesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
