% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{getVisualPropertyDefault}
\alias{getVisualPropertyDefault}
\title{Get Visual Property Default}
\usage{
getVisualPropertyDefault(property, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{property}{Name of property, e.g., NODE_FILL_COLOR (see \link{getVisualPropertyNames})}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Retrieve the default value for a visual property.
}
\examples{
\donttest{
getVisualPropertyDefault('NODE_SIZE')
}
}
