% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getVisualPropertyNames}
\alias{getVisualPropertyNames}
\title{Get Visual Property Names}
\usage{
getVisualPropertyNames(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getVisualPropertyNames()
}
}
