% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getVisualStyleJSON}
\alias{getVisualStyleJSON}
\title{Get Visual Style JSON}
\usage{
getVisualStyleJSON(styleName, css = FALSE, base.url = .defaultBaseUrl)
}
\arguments{
\item{styleName}{name of style}

\item{css}{TRUE to create a CytoscapeJS CSS style, FALSE to create 
a generic JSON version. Default is FALSE.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of visual style properties
}
\description{
Get all defaults and mappings for a visual style
}
\examples{
\donttest{
getVisualStyleJSON()
}
}
