% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getWindowID}
\alias{getWindowID}
\title{DEPRECATED: getWindowID}
\usage{
Use the replacement function instead:
\code{\link[RCy3]{getNetworkSuid}}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
