% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getWindowList}
\alias{getWindowList}
\title{DEPRECATED: getWindowList}
\usage{
Use the replacement function instead:
\link{getNetworkList}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
