% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{importFilters}
\alias{importFilters}
\title{Import Filters}
\usage{
importFilters(filename, base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) Path and name of the filters file to load.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Path to saved file
}
\description{
Loads filters from a file in JSON format.
}
\examples{
\donttest{
importFilters()
}
}
