% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{importNetworkFromFile}
\alias{importNetworkFromFile}
\title{Import Network From File}
\usage{
importNetworkFromFile(file = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{file}{Name of file in any of the supported formats (e.g., SIF, GML, XGMML, etc).
If NULL, a demo network file in SIF format is loaded.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
(int) network SUID
}
\description{
Loads a network from specified file
}
\examples{
\donttest{
importNetworkFromFile()
}
}
