% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{importVisualStyles}
\alias{importVisualStyles}
\title{Import Visual Styles}
\usage{
importVisualStyles(filename = "styles.xml", base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) Name of the style file to load. Only reads XML files.
Default is "styles.xml".}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
(list) Names of styles loaded
}
\description{
Loads styles from an XML file and returns the names of the 
loaded styles.
}
\examples{
\donttest{
importVisualStyles()
}
}
\seealso{
exportVisualStyles
}
