% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{initEdgeAttribute}
\alias{initEdgeAttribute}
\alias{initEdgeAttributedefunct}
\title{DEFUNCT: initEdgeAttribute}
\usage{
initEdgeAttributedefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
