% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{initNodeAttribute}
\alias{initNodeAttribute}
\alias{initNodeAttributedefunct}
\title{DEFUNCT: initNodeAttribute}
\usage{
initNodeAttributedefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
