% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\docType{data}
\name{JupyterBRIDGEURL}
\alias{JupyterBRIDGEURL}
\title{getJupyterBridgeURL}
\format{
An object of class \code{character} of length 1.
}
\usage{
JupyterBRIDGEURL
}
\value{
jupyter bridge server url
}
\description{
Get the jupyter bridge server url
}
\examples{
\donttest{
getJupyterBridgeURL()
}
}
\keyword{datasets}
