% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{listGroups}
\alias{listGroups}
\title{List Groups}
\usage{
listGroups(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of group SUIDs
}
\description{
Retrieve a list of all group SUIDs in a network.
}
\examples{
\donttest{
listGroups()
}
}
