% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{listTableColumns}
\alias{listTableColumns}
\title{List names of all columns in a table}
\usage{
listTableColumns(table = "node", namespace = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{name of table, e.g., node, edge, network; default is "node"}

\item{namespace}{namespace of table, e.g., default}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of column names
}
\description{
List names of all columns in a table
}
\examples{
\donttest{
listTableColumns()
listTableColumns('edge')
listTableColumns('network')
}
}
