% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{makeRandomGraph}
\alias{makeRandomGraph}
\alias{makeRandomGraphdefunct}
\title{DEFUNCT: makeRandomGraph}
\usage{
makeRandomGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
