% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{makeSimpleIgraph}
\alias{makeSimpleIgraph}
\title{Make Simple Igraph}
\usage{
makeSimpleIgraph()
}
\value{
A \code{igraph} object with a few nodes, edges and attributes
}
\description{
This function creates a simple iGraph object with various
node and edge attribute types to help demonstrate round trip conversion with
Cytoscape networks via RCy3.
}
\examples{
{
makeSimpleIgraph()
}
}
\seealso{
createNetworkFromIgraph, createIgraphFromNetwork, makeSimpleGraph
}
