% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{matchArrowColorToEdge}
\alias{matchArrowColorToEdge}
\title{Match Arrow Color To Edge}
\usage{
matchArrowColorToEdge(new.state, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.state}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
matchArrowColorToEdge()
}
}
