% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{noa.names}
\alias{noa.names}
\alias{noa.namesdefunct}
\title{DEFUNCT: noa.names}
\usage{
noa.namesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
