% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{nodeSet2JSON}
\alias{nodeSet2JSON}
\title{Creates a table of nodes to CyREST JSON}
\usage{
nodeSet2JSON(node.set, node.id.list = "id", ...)
}
\arguments{
\item{node.set}{(data.frame) each row is a node and columns contain node attributes}

\item{node.id.list}{(char) override default list name for node ids
Adapted from Ruth Isserlin's CellCellINteractions_utility_functions.R}
}
\description{
Creates a table of nodes to CyREST JSON
}
