% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\docType{data}
\name{notebookIsRunning}
\alias{notebookIsRunning}
\title{setNotebookIsRunning}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
notebookIsRunning
}
\description{
setNotebookIsRunning
}
\examples{
\donttest{
setNotebookIsRunning()
}
}
\keyword{datasets}
