% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterface.R
\name{openCySwagger}
\alias{openCySwagger}
\title{Open CySwagger docs in browser}
\usage{
openCySwagger(domain = "cyrest", obj = CytoscapeConnection())
}
\arguments{
\item{domain}{(char) documentation domain or scope}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
Web page
}
\description{
Opens swagger docs in default browser for a live
instance of CyREST or CyREST-supported operations.
}
\examples{
\donttest{
openCySwagger()
openCySwagger('commands')
}
}
