% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{paletteColorBrewerPuBu}
\alias{paletteColorBrewerPuBu}
\title{paletteColorBrewerPuBu Sequential}
\usage{
paletteColorBrewerPuBu(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 9
. See RColorBrewer::display.brewer.all()}
}
\value{
List of palette colors
}
\description{
Generate a sequential PuBu Brewer palette of a given size
}
\examples{
paletteColorBrewerPuBu()
}
