% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{paletteColorBrewerSet2}
\alias{paletteColorBrewerSet2}
\title{paletteColorBrewerSet2 Qualitative}
\usage{
paletteColorBrewerSet2(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 8
. See RColorBrewer::display.brewer.all()}
}
\value{
List of palette colors
}
\description{
Generate a qualitative Set2 Brewer palette of a given size
}
\examples{
\donttest{
paletteColorBrewerSet2()
}
}
