% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{paletteColorRandom}
\alias{paletteColorRandom}
\title{paletteColorRandom Qualitative}
\usage{
paletteColorRandom(value.count = 1)
}
\arguments{
\item{value.count}{Number of colors to generate; default is 1}
}
\value{
List of random colors
}
\description{
Generate random color map of a given size
}
\examples{
\donttest{
paletteColorRandom()
}
}
\seealso{
genNodeColorMap genEdgeColorMap
}
