% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{paletteColorVirdis}
\alias{paletteColorVirdis}
\title{paletteColorVirdis}
\usage{
paletteColorVirdis(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 9
Get colors from https://cran.r-project.org/web/packages/viridis/index.html
. See virdis::virdis()}
}
\value{
List of palette colors
}
\description{
Generate a Virdis Brewer palette of a given size
}
\examples{
paletteColorVirdis()
}
