% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{predictTimeToDisplayGraph}
\alias{predictTimeToDisplayGraph}
\alias{predictTimeToDisplayGraphdefunct}
\title{DEFUNCT: predictTimeToDisplayGraph}
\usage{
predictTimeToDisplayGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
