% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{removeNodeCustomChart}
\alias{removeNodeCustomChart}
\title{Remove Node Custom Chart}
\usage{
removeNodeCustomChart(slot = 1, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{slot}{(optional) Which custom graphics slot to modify. Slots 1-9 are 
available for independent charts and images. Default is 1.}

\item{style.name}{(optional) Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Remove the default custom graphics.
}
\examples{
\donttest{
removeNodeCustomChart()
}
}
