% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{renameCytoscapeNetwork,CytoscapeWindowClass-method}
\alias{renameCytoscapeNetwork,CytoscapeWindowClass-method}
\title{Rename a network}
\usage{
\S4method{renameCytoscapeNetwork}{CytoscapeWindowClass}(obj, new_title,
  copy.graph.to.R = FALSE)
}
\arguments{
\item{new_title}{New name for the copy}

\item{copy.graph.to.R}{Logical whether to copy the graph to a new object in R}

\item{object}{Cytoscape network}
}
\value{
Connection to the renamed network.
}
\description{
Renames a Cytoscape Network.
}
\examples{
  dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
renamed_net <- renameCytoscapeNetwork(cw, "renamed_network")

}
}
\keyword{graph}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\seealso{
\code{\link{createWindowFromSelection}}, \code{\link{existing.CytoscapeWindow}}, \code{\link{copyCytoscapeNetwork}}
}
\concept{
RCy3
}

