% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{renameNetwork,OptionalCyWinClass-method}
\alias{renameNetwork,OptionalCyWinClass-method}
\title{Rename a network}
\usage{

  \S4method{renameNetwork}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork(),
  new.title, return.graph = FALSE)
}
\arguments{
\item{new.title}{New name for the copy}

\item{return.graph}{Logical whether to copy the graph to a new object in R}

\item{object}{Cytoscape network}
}
\value{
Connection to the renamed network.
}
\description{
Renames a Cytoscape Network.
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
renamed_net <- renameNetwork(cw, "renamed_network")
}

}
\seealso{
\code{\link{createNetworkFromSelection}}, \code{\link{CytoscapeWindowFromNetwork}}, \code{\link{cloneNetwork}}
}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\concept{
RCy3
}
