% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename a network}
\usage{
renameNetwork(new.title, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{new.title}{New name for the network}

\item{network}{name or suid of the network that you want to rename; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
None
}
\description{
Sets a new name for this network
}
\details{
Duplicate network names are not allowed
}
\examples{
\dontrun{
renameNetwork("renamed network")
}
}
\author{
Alexander Pico, Julia Gustavsen
}
