% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{renameTableColumn}
\alias{renameTableColumn}
\title{Rename Table Column}
\usage{
renameTableColumn(column, new.name, table = "node", namespace = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{column}{Name of the column to rename}

\item{new.name}{New name for the specified column}

\item{table}{(optional) Name of table, e.g., node (default), edge or network}

\item{namespace}{(optional) Namespace of table, e.g., default (default), shared
or hidden}

\item{network}{(optional) Name or SUID of the network. Default is the "current" 
network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Sets a new name for a column.
}
\examples{
\donttest{
renameTableColumn('exp','log2FC')
}
}
