% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{resetDefaultSandbox}
\alias{resetDefaultSandbox}
\title{resetDefaultSandbox}
\usage{
resetDefaultSandbox()
}
\description{
Reset the entire state of the sandbox system.
}
\examples{
\donttest{
resetDefaultSandbox()
}
}
