% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{runningRemoteCheck}
\alias{runningRemoteCheck}
\title{runningRmoteCheck}
\usage{
runningRemoteCheck(base.url = .defaultBaseUrl, newState = NULL)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}

\item{newState}{new state of running remote}
}
\value{
oldState
}
\description{
runningRmoteCheck
}
\examples{
\donttest{
runningRmoteCheck()
}
}
