% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxGetFileInfo}
\alias{sandboxGetFileInfo}
\title{sandboxGetFileInfo}
\usage{
sandboxGetFileInfo(fileName, sandboxName = NULL, base.url = .defaultBaseUrl)
}
\value{
sandboxGetFileInfo
}
\description{
sandboxGetFileInfo
}
\examples{
\donttest{
sandboxGetFileInfo()
}
}
