% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{sandboxInitializer}
\alias{sandboxInitializer}
\title{sandboxInitializer}
\usage{
sandboxInitializer(newSandbox = NULL, ...)
}
\value{
sandbox
}
\description{
sandboxInitializer.
}
\examples{
\donttest{
sandboxInitializer()
}
}
