% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxSendTo}
\alias{sandboxSendTo}
\title{sandboxSendTo}
\usage{
sandboxSendTo(
  sourceFile,
  destFile = NULL,
  overwrite = TRUE,
  sandboxName = NULL,
  base.url = base.url
)
}
\value{
sandboxSendTo
}
\description{
sandboxSendTo
}
\examples{
\donttest{
sandboxSendTo()
}
}
