% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{saveLayout}
\alias{saveLayout}
\alias{saveLayoutdefunct}
\title{DEFUNCT: saveLayout}
\usage{
saveLayoutdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{layoutCopycat}
}
