% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{saveNetwork,OptionalCyWinClass-method}
\alias{saveNetwork,OptionalCyWinClass-method}
\title{Save a network in one of mulitple file formats}
\usage{

  \S4method{saveNetwork}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork(),
  filename, type = "cys")
}
\description{
Save a network in one of mulitple file formats
}
