% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{saveNetwork}
\alias{saveNetwork}
\title{Save a network in one of mulitple file formats}
\usage{
saveNetwork(filename, type = "cys", base.url = .defaultBaseUrl)
}
\description{
Save a network in one of mulitple file formats
}
