% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.R
\name{saveSession}
\alias{saveSession}
\title{Save Session to File}
\usage{
saveSession(filename = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) name of the session file to save}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Saves the current Cytoscape session as a CYS file.
}
\details{
If no \code{filename} is provided, then it attempts to
save to an existing CYS file associated with the session. If 
\code{filename} already exists, then it is overwritten.
}
\examples{
\donttest{
saveSession('myFirstSession')
saveSession() 
}
}
