% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{saveVisualStyle}
\alias{saveVisualStyle}
\title{Saves the current visual style as a data file}
\usage{
saveVisualStyle(filename, type, obj = CytoscapeConnection())
}
\arguments{
\item{filename}{(char) name of the style file to save}

\item{type}{(char) type of data file to export, e.g., XML, JSON (case sensitive)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Saves the current visual style as a data file
}
\examples{
\donttest{
saveVisualStyle('myStyle','JSON')
}
}
