% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Style.R
\name{saveVisualStyles}
\alias{saveVisualStyles}
\title{Save Visual Styles}
\usage{
saveVisualStyles(filename = NULL, type = NULL, styles = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{filename}{(char) name of the style file to save. Default is "styles.xml"}

\item{type}{(char) type of data file to export, e.g., XML, JSON (case sensitive). Default is XML.}

\item{styles}{(optional) The styles to be exported, listed as a comma-separated string. 
If no styles are specified, only the current one is exported.}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Saves one or more visual styles to file
}
\examples{
\donttest{
saveVisualStyles()
saveVisualStyles('myStyle', type = 'JSON')
saveVisualStyles('myStyle', style = 'Minimal,default,Directed')
}
}
