% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{schemeColorBrewerAccent}
\alias{schemeColorBrewerAccent}
\title{schemeColorBrewerAccent Qualitative}
\usage{
schemeColorBrewerAccent(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 8
. See RColorBrewer::display.brewer.all()}
}
\value{
List of palette colors
}
\description{
Generate Accent Brewer palette of a given size
}
\examples{
\donttest{
schemeColorBrewerAccent()
}
}
