% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{schemeColorBrewerPaired}
\alias{schemeColorBrewerPaired}
\title{schemeColorBrewerPaired Qualitative}
\usage{
schemeColorBrewerPaired(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 
12. See RColorBrewer::display.brewer.all()}
}
\value{
List of palette colors
}
\description{
Generate Paired Brewer palette of a given size
}
\examples{
\donttest{
schemeColorBrewerPaired()
}
}
