% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{schemeColorBrewerSet3}
\alias{schemeColorBrewerSet3}
\title{schemeColorBrewerSet3 Qualitative}
\usage{
schemeColorBrewerSet3(value.count = 3)
}
\arguments{
\item{value.count}{Number of colors to generate; min is 3 (default); max is 
12. See RColorBrewer::display.brewer.all()}
}
\value{
List of palette colors
}
\description{
Generate Set3 Brewer palette of a given size
}
\examples{
\donttest{
schemeColorBrewerSet3()
}
}
