% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{schemeColorRandom}
\alias{schemeColorRandom}
\title{schemeColorRandom Qualitative}
\usage{
schemeColorRandom(value.count = 1)
}
\arguments{
\item{value.count}{Number of colors to generate; default is 1}
}
\value{
List of random colors
}
\description{
Generate random color map of a given size
}
\examples{
\donttest{
schemeColorRandom()
}
}
\seealso{
genNodeColorMap genEdgeColorMap
}
