% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{selectAllEdges,CytoscapeWindowClass-method}
\alias{selectAllEdges,CytoscapeWindowClass-method}
\title{Select all edges}
\usage{
\S4method{selectAllEdges}{CytoscapeWindowClass}(obj)
}
\arguments{
\item{object}{Cytoscape network}
}
\value{
Selects all edges in a specified network.
}
\description{
Selects all edges in a Cytoscape Network
}
\examples{
  \dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllEdges(cw)

}
}
\keyword{graph}

\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\seealso{
\code{\link{selectEdges}}
}
\concept{
RCy3
}

