% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{selectAllEdges,OptionalCyWinClass-method}
\alias{selectAllEdges,OptionalCyWinClass-method}
\title{Select all edges}
\usage{

  \S4method{selectAllEdges}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork())
}
\arguments{
\item{obj}{(optional) \code{CytoscapeConnection} or \code{CytoscapeWindow}}
}
\value{
Selects all edges in a specified network.
}
\description{
Selects all edges in a Cytoscape Network
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllEdges(cw)
}

}
\seealso{
\code{\link{selectEdges}}
}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\concept{
RCy3
}
