% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{selectAllEdges}
\alias{selectAllEdges}
\title{Select all edges}
\usage{
selectAllEdges(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
Selects all edges in a specified network.
}
\description{
Selects all edges in a Cytoscape Network
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllEdges(cw)
}
}
\author{
Alexander Pico, Julia Gustavsen
}
