% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{selectAllNodes,CytoscapeWindowClass-method}
\alias{selectAllNodes,CytoscapeWindowClass-method}
\title{Select all nodes}
\usage{
\S4method{selectAllNodes}{CytoscapeWindowClass}(obj)
}
\arguments{
\item{object}{Cytoscape network}
}
\value{
Selects all nodes in a specified network.
}
\description{
Selects all nodes in a Cytoscape Network
}
\examples{
  dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllNodes(cw)

}
}
\keyword{graph}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\seealso{
\code{\link{selectNodes}}
}
\concept{
RCy3
}

