% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{selectAllNodes,OptionalCyWinClass-method}
\alias{selectAllNodes,OptionalCyWinClass-method}
\title{Select all nodes}
\usage{

  \S4method{selectAllNodes}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork())
}
\arguments{
\item{obj}{(optional) \code{CytoscapeConnection} or \code{CytoscapeWindow}}
}
\value{
Selects all nodes in a specified network.
}
\description{
Selects all nodes in a Cytoscape Network
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllNodes(cw)
}

}
\seealso{
\code{\link{selectNodes}}
}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\concept{
RCy3
}
