% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{selectAllNodes}
\alias{selectAllNodes}
\title{Select all nodes}
\usage{
selectAllNodes(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network into which you want to select; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
Selects all nodes in a specified network.
}
\description{
Selects all nodes in a Cytoscape Network
}
\examples{
\donttest{
selectAllNodes()
}
}
\seealso{
\code{\link{selectNodes}}
}
\author{
Alexander Pico, Julia Gustavsen
}
