% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{selectEdgesConnectedBySelectedNodes}
\alias{selectEdgesConnectedBySelectedNodes}
\title{Select the edges connecting selected nodes in Cytoscape Network}
\usage{
selectEdgesConnectedBySelectedNodes(obj)
}
\arguments{
\item{object}{Cytoscape network}
}
\value{
network with edges selected
}
\description{
Selects edges in a Cytoscape Network connecting the selected nodes
}
\examples{
\dontrun{
cw <- CytoscapeWindow('vignette select edges', graph = RCy3::makeSimpleGraph(), overwrite = TRUE)
displayGraph(cw)
selectNodes(cw,"A") # selects specific nodes
getSelectedNodes(cw)
getSelectedEdges(cw)
selectFirstNeighborsOfSelectedNodes(cw)
## This has only selected the nodes, but not the edges in Cytoscape, so we will need to select all of the edges before we make the new subnetwork.
selectEdgesConnectedBySelectedNodes(cw)
getSelectedNodes(cw)
getSelectedEdges(cw)
}

}
\author{
Julia Gustavsen, \email{j.gustavsen@gmail.com}
}
\seealso{
\code{\link{createWindowFromSelection}}, \code{\link{selectEdgesConnectedBySelectedNodes}}, \code{\link{renameCytoscapeNetwork}}
}
\concept{
RCy3
}

