% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{selectEdgesConnectingSelectedNodes}
\alias{selectEdgesConnectingSelectedNodes}
\title{Select the edges connecting selected nodes in Cytoscape Network}
\usage{
selectEdgesConnectingSelectedNodes(network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
network with edges selected
}
\description{
Selects edges in a Cytoscape Network connecting the selected nodes
}
\examples{
\dontrun{
}

}
\author{
Alexander Pico, Julia Gustavsen
}
