% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{sendEdges}
\alias{sendEdges}
\title{DEPRECATED: sendEdges}
\usage{
Use the replacement function instead:
\code{\link[RCy3]{sendEdgesFromGraph}}
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
