% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setCommandProperties}
\alias{setCommandProperties}
\alias{setCommandPropertiesdefunct}
\title{DEFUNCT: setCommandProperties}
\usage{
setCommandPropertiesdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{commandsRun}
}
