% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{setCurrentNetwork}
\alias{setCurrentNetwork}
\title{Set current network}
\usage{
setCurrentNetwork(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network that you want set as current}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
server response
}
\description{
Selects the given network as "current"
}
\examples{
\donttest{
setCurrentNetwork('MyNetwork')
}
}
