% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{setCurrentView}
\alias{setCurrentView}
\title{Set Current View}
\usage{
setCurrentView(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
Takes first (presumably only) view associated with provided network
}
\examples{
\donttest{
setCurrentView()
}
}
