% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setDefaultEdgeReverseSelectionColor}
\alias{setDefaultEdgeReverseSelectionColor}
\alias{setDefaultEdgeReverseSelectionColordefunct}
\title{DEFUNCT: setDefaultEdgeReverseSelectionColor}
\usage{
setDefaultEdgeReverseSelectionColordefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
