% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setDefaultNodeReverseSelectionColor}
\alias{setDefaultNodeReverseSelectionColor}
\alias{setDefaultNodeReverseSelectionColordefunct}
\title{DEFUNCT: setDefaultNodeReverseSelectionColor}
\usage{
setDefaultNodeReverseSelectionColordefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
