% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{setDefaultSandboxPath}
\alias{setDefaultSandboxPath}
\title{setDefaultSandboxPath}
\usage{
setDefaultSandboxPath(newPath)
}
\arguments{
\item{newPath}{new path of default sandbox}
}
\value{
default sandbox path
}
\description{
Set and return the default path, which isn't one of the properties tracked in the default_sandbox.
}
\examples{
\donttest{
setDefaultSandboxPath()
}
}
