% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setEdgeAttributes}
\alias{setEdgeAttributes}
\alias{setEdgeAttributesdefunct}
\title{DEFUNCT: setEdgeAttributes}
\usage{
setEdgeAttributesdefunct
}
\value{
None
}
\description{
This function is defunct and removed in the next releases.
}
