% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setEdgeLabelOpacityDefault}
\alias{setEdgeLabelOpacityDefault}
\title{Set Edge Label Opacity Default}
\usage{
setEdgeLabelOpacityDefault(
  new.opacity,
  style.name = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{new.opacity}{Numeric values between 0 and 255; 0 is invisible.}

\item{style.name}{Name of style; default is "default" style.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set default opacity value for all unmapped edge labels.
}
\examples{
\donttest{
setEdgeLabelOpacityDefault(50)
}
}
