% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setEdgeLineStyleDefault}
\alias{setEdgeLineStyleDefault}
\title{Set Edge Line Style Default}
\usage{
setEdgeLineStyleDefault(new.line.style, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.line.style}{Name of line style, e.g., SOLID, LONG_DASH, etc (see \link{getLineStyles})}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default edge style.
}
\examples{
\donttest{
setEdgeLineStyleDefault('LONG_DASH')
}
}
