% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeLineStyleMapping}
\alias{setEdgeLineStyleMapping}
\title{Set Edge Line Style Mapping}
\usage{
setEdgeLineStyleMapping(table.column, table.column.values, line.styles,
  default.line.style = "SOLID", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{line.styles}{DESCRIPTION}

\item{default.line.style}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeLineStyleMapping()
}
}
