% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setEdgeLineWidthDefault}
\alias{setEdgeLineWidthDefault}
\title{Set Edge Line Width Default}
\usage{
setEdgeLineWidthDefault(new.width, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.width}{Numeric value for width}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default edge width.
}
\examples{
\donttest{
setEdgeLineWidthDefault(3)
}
}
