% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeLineWidthMapping}
\alias{setEdgeLineWidthMapping}
\title{Set Edge Line Width Mapping}
\usage{
setEdgeLineWidthMapping(table.column, table.column.values = NULL,
  widths = NULL, mapping.type = "c", default.width = NULL,
  style.name = "default", network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{widths}{DESCRIPTION}

\item{mapping.type}{(char) continuous, discrete or passthrough (c,d,p); default is continuous}

\item{default.width}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeLineWidthMapping()
}
}
