% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeOpacityMapping}
\alias{setEdgeOpacityMapping}
\title{Set Edge Opacity Mapping}
\usage{
setEdgeOpacityMapping(table.column, table.column.values = NULL,
  opacities = NULL, mapping.type = "c", default.opacity = NULL,
  style.name = "default", network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{opacities}{(integer) values between 0 and 255; 0 is invisible}

\item{mapping.type}{(char) continuous, discrete or passthrough (c,d,p); default is continuous}

\item{default.opacity}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeOpacityMapping()
}
}
