% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDirects.R
\name{setEdgePropertyDirect}
\alias{setEdgePropertyDirect}
\title{Set Edge Property Direct}
\usage{
setEdgePropertyDirect(edge.names, new.values, visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{DESCRIPTION}

\item{new.values}{DESCRIPTION}

\item{visual.property}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgePropertyDirect()
}
}
