% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeSourceArrowColorMapping}
\alias{setEdgeSourceArrowColorMapping}
\title{Set Edge Source Arrow Color Mapping}
\usage{
setEdgeSourceArrowColorMapping(edge.attribute.name, table.column.values, colors,
  mode = "interpolate", default.color = "#000000", style.name = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.attribute.name}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{colors}{DESCRIPTION}

\item{mode}{DESCRIPTION}

\item{default.color}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeSourceArrowColorMapping()
}
}
