% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeTargetArrowColorMapping}
\alias{setEdgeTargetArrowColorMapping}
\title{Set Edge Target Arrow Color Mapping}
\usage{
setEdgeTargetArrowColorMapping(table.column, table.column.values = NULL,
  colors = NULL, mapping.type = "c", default.color = NULL,
  style.name = "default", network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{colors}{DESCRIPTION}

\item{mapping.type}{(char) continuous, discrete or passthrough (c,d,p); default is continuous}

\item{default.color}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeTargetArrowColorMapping()
}
}
