% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeTargetArrowMapping}
\alias{setEdgeTargetArrowMapping}
\title{Set Edge Target Arrow Mapping}
\usage{
setEdgeTargetArrowMapping(edge.attribute.name, attribute.values, arrows,
  default = "ARROW", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.attribute.name}{DESCRIPTION}

\item{attribute.values}{DESCRIPTION}

\item{arrows}{DESCRIPTION}

\item{default}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeTargetArrowMapping()
}
}
