% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeTooltipMapping}
\alias{setEdgeTooltipMapping}
\title{Set Edge Tooltip Mapping}
\usage{
setEdgeTooltipMapping(edge.attribute.name, style.name = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.attribute.name}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeTooltipMapping()
}
}
