% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setGraph}
\alias{setGraph}
\alias{setGraphdefunct}
\title{DEFUNCT: setGraph}
\usage{
setGraphdefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
