% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\name{setModelPropagationSecs}
\alias{setModelPropagationSecs}
\title{Set Model Propagation Delay}
\usage{
setModelPropagationSecs(secs = 5)
}
\arguments{
\item{secs}{Number of seconds to delay.}
}
\value{
None
}
\description{
This function sets an internal delay variable that allows
Cytoscape to "catchup" prior to subsequent functions. Call without specifying
\code{secs} to restore default value.
}
\details{
This delay is only necessary while concurrency bugs exist in the
Cytoscape application. This delay may need to be increased from the default
value in certain use cases, e.g., larger networks.
}
\examples{
{
setModelPropagationSecs(2)
setModelPropagationSecs() #restores default delay
}
}
\seealso{
setCatchupFilterSecs, setCatchupNetworkSecs
}
