% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{setNetworkCenterBypass}
\alias{setNetworkCenterBypass}
\title{Set Network Center Bypass}
\usage{
setNetworkCenterBypass(x, y, bypass = FALSE, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{x}{Coordinate value, increases going to the right.}

\item{y}{Coordinate value, increase going down.}

\item{bypass}{Whether to set permanent bypass value. Default is \code{FALSE} 
per common use of temporary center settings.}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the bypass value for center x and y for the network. This
function could be used to pan and scroll the Cytoscape canvas.
}
\details{
This method permanently overrides any default values 
for this visual property. This method
ultimately calls the generic function, \link{setNetworkPropertyBypass}, which 
can be used to set any visual property. To restore defaults, use
 \link{clearNetworkPropertyBypass}.
}
\examples{
\donttest{
setNetworkCenterBypass()
}
}
\seealso{
{
\link{setNetworkPropertyBypass},
\link{clearNetworkPropertyBypass}
}
}
