% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setNodeAttributes}
\alias{setNodeAttributes}
\alias{setNodeAttributesdefunct}
\title{DEFUNCT: setNodeAttributes}
\usage{
setNodeAttributesdefunct
}
\value{
None
}
\description{
This function is defunct and removed in the next releases.
}
