% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeBorderColorMapping}
\alias{setNodeBorderColorMapping}
\title{Set Node Border Color Mapping}
\usage{
setNodeBorderColorMapping(table.column, table.column.values, colors, mode,
  default.color = "#000000", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{colors}{DESCRIPTION}

\item{mode}{DESCRIPTION}

\item{default.color}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeBorderColorMapping()
}
}
