% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeBorderWidthMapping}
\alias{setNodeBorderWidthMapping}
\title{Set Node Border Width Mapping}
\usage{
setNodeBorderWidthMapping(table.column, attribute.values, line.widths,
  default.width = 1, style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{attribute.values}{DESCRIPTION}

\item{line.widths}{DESCRIPTION}

\item{default.width}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeBorderWidthMapping()
}
}
