% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeColorDefault}
\alias{setNodeColorDefault}
\title{Set Node Color Default}
\usage{
setNodeColorDefault(new.color, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.color}{Color as hex code, e.g., #FD5903}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default node color.
}
\examples{
\donttest{
setNodeColorDefault('#FD5903')
}
}
