% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeColorMapping}
\alias{setNodeColorMapping}
\title{Set Node Color Mapping}
\usage{
setNodeColorMapping(table.column, table.column.values, colors,
  mapping.type = "c", default.color = NULL, style.name = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{colors}{DESCRIPTION}

\item{mapping.type}{(char) continuous, discrete or passthrough (c,d,p)}

\item{default.color}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeColorMapping()
}
}
