% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeCustomPosition}
\alias{setNodeCustomPosition}
\title{Set Node Custom Position}
\usage{
setNodeCustomPosition(
  nodeAnchor = "C",
  graphicAnchor = "C",
  justification = "c",
  xOffset = 0,
  yOffset = 0,
  slot = 1,
  style.name = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{nodeAnchor}{Position on node to place the graphic: NW,N,NE,E,SE,S,SW,W 
or C for center (default)}

\item{graphicAnchor}{Position on graphic to place on node: NW,N,NE,E,SE,S,SW,W 
or C for center (default)}

\item{justification}{Positioning of content within graphic: l,r,c (default)}

\item{xOffset}{Additional offset in the x direction}

\item{yOffset}{Additional offset in the y direction}

\item{slot}{(optional) Which custom graphics slot to modify. Slots 1-9 are 
available for independent charts, gradients and images. Default is 1.}

\item{style.name}{(optional) Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Adjust the position of a custom graphic relative to its node.
}
\examples{
\donttest{
setNodeCustomPosition()
}
}
