% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{setNodeFillOpacityBypass}
\alias{setNodeFillOpacityBypass}
\title{Set Node Fill Opacity Bypass}
\usage{
setNodeFillOpacityBypass(node.names, new.values, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{DESCRIPTION}

\item{new.values}{DESCRIPTION}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
FUNCTION_DESCRIPTION
}
\details{
This method permanently overrides any default values or mappings 
defined for this visual property of the node or nodes specified. This method
ultimately calls the generic function, \link{setNodePropertyBypass}, which 
can be used to set any visual property. To restore defaults and mappings, use
 \link{clearNodePropertyBypass}.
}
\examples{
\donttest{
setNodeFillOpacityBypass()
}
}
\seealso{
{
\link{setNodePropertyBypass},
\link{clearNodePropertyBypass}
}
}
