% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeFontFaceDefault}
\alias{setNodeFontFaceDefault}
\title{Set Node Font Face Default}
\usage{
setNodeFontFaceDefault(new.font, style.name = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{new.font}{String specification of font face, style and size, e.g., 
"SansSerif,plain,12" or "Dialog,plain,10"}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default node font.
}
\examples{
\donttest{
setNodeFontFaceDefault("Dialog,plain,10")
}
}
