% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeLabelDefault}
\alias{setNodeLabelDefault}
\title{Set Node Label Default}
\usage{
setNodeLabelDefault(new.label, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.label}{String label for unmapped nodes.}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default node label.
}
\examples{
\donttest{
setNodeLabelDefault('unknown')
}
}
