% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeLabelMapping}
\alias{setNodeLabelMapping}
\title{Set Node Label Mapping}
\usage{
setNodeLabelMapping(table.column, style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeLabelMapping()
}
}
