% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeLabelOpacityDefault}
\alias{setNodeLabelOpacityDefault}
\title{Set Node Label Opacity Default}
\usage{
setNodeLabelOpacityDefault(new.opacity, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.opacity}{Numeric values between 0 and 255; 0 is invisible.}

\item{style.name}{Name of style; default is "default" style.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set default opacity value for all unmapped node labels.
}
\examples{
\donttest{
setNodeLabelOpacityDefault(50)
}
}
