% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeLabelPositionDefault}
\alias{setNodeLabelPositionDefault}
\title{Set Node Label Position Default}
\usage{
setNodeLabelPositionDefault(
  new.nodeAnchor,
  new.graphicAnchor,
  new.justification,
  new.xOffset,
  new.yOffset,
  style.name = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{new.nodeAnchor}{New position on node to place the graphic: C,NW,N,NE,E,SE,S,SW,W}

\item{new.graphicAnchor}{New position on graphic to place on node: C,NW,N,NE,E,SE,S,SW,W}

\item{new.justification}{New positioning of content within graphic: l,r,c}

\item{new.xOffset}{New additional offset in the x direction}

\item{new.yOffset}{New additional offset in the y direction}

\item{style.name}{Name of style; default is "default" style.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default node label position
}
\examples{
\donttest{
setNodeLabelPositionDefault("S","C","c",0.00,0.00)
}
}
