% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDirects.R
\name{setNodePropertyDirect}
\alias{setNodePropertyDirect}
\title{Set Node Property Direct}
\usage{
setNodePropertyDirect(node.names, new.values, visual.property, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{node.names}{DESCRIPTION}

\item{new.values}{DESCRIPTION}

\item{visual.property}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodePropertyDirect()
}
}
