% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeShapeMapping}
\alias{setNodeShapeMapping}
\title{Set Node Shape Mapping}
\usage{
setNodeShapeMapping(table.column, attribute.values, node.shapes,
  default.shape = "ELLIPSE", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{attribute.values}{DESCRIPTION}

\item{node.shapes}{DESCRIPTION}

\item{default.shape}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeShapeMapping()
}
}
