% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{setNotebookIsRunning}
\alias{setNotebookIsRunning}
\title{setNotebookIsRunning}
\usage{
setNotebookIsRunning(newState = NULL)
}
\description{
setNotebookIsRunning
}
\examples{
\donttest{
setNotebookIsRunning()
}
}
