% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setTooltipInitialDelay}
\alias{setTooltipInitialDelay}
\alias{setTooltipInitialDelaydefunct}
\title{DEFUNCT: setTooltipInitialDelay}
\usage{
setTooltipInitialDelaydefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
