% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setVisualPropertyDefault}
\alias{setVisualPropertyDefault}
\title{Set Visual Property Default}
\usage{
setVisualPropertyDefault(style.string, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{style.string}{A named list including "visualProperty" and "value"}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set the default value for a visual property.
}
\examples{
\donttest{
setVisualPropertyDefault(list(visualProperty = "NODE_SIZE", value = 35))
}
}
