% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{setVisualStyle}
\alias{setVisualStyle}
\title{Set Visual Style}
\usage{
setVisualStyle(style.name, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
Apply a visual style to a network
}
\examples{
\donttest{
setVisualStyle()
}
}
