% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{setWindowSize}
\alias{setWindowSize}
\alias{setWindowSizedefunct}
\title{DEFUNCT: setWindowSize}
\usage{
setWindowSizedefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
}
