% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{showImageInNotebook}
\alias{showImageInNotebook}
\title{Show Image in Notebook}
\usage{
showImageInNotebook(filename)
}
\arguments{
\item{filename}{(\code{character}) Full path or path relavtive to current 
working directory, in addition to the name of the file. Extension is 
automatically added based on the \code{type} argument. If blank, the current
network name will be used.}
}
\value{
None
}
\description{
Show Image in Notebook
}
\details{
Show Image in Notebook
}
\examples{
\donttest{
showImageInNotebook()
}
}
