% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{syncNodeCustomGraphicsSize}
\alias{syncNodeCustomGraphicsSize}
\title{Sync Node Custom Graphics Size}
\usage{
syncNodeCustomGraphicsSize(new.state, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.state}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
syncNodeCustomGraphicsSize()
}
}
