% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{toggleGraphicsDetails}
\alias{toggleGraphicsDetails}
\title{Toggle Graphics Details}
\usage{
toggleGraphicsDetails(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
None
}
\description{
Regardless of the current zoom level and network size,
show (or hide) graphics details, e.g., node labels.
}
\details{
Displaying graphics details on a very large network will affect pan
and zoom performance, depending on your available RAM. 
See \link{cytoscapeMemoryStatus}.
}
\examples{
\donttest{
showGraphicsDetails(TRUE)
}
}
