% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{unhideEdges}
\alias{unhideEdges}
\title{Unhide Edges}
\usage{
unhideEdges(edge.names, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{List of edge names}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Unhide specified edges that were previously hidden, by 
clearing the Visible property bypass value.
}
\details{
This method permanently overrides any default values or mappings 
defined for this visual property of the edge or edges specified. This method
ultimately calls the generic function, \link{clearEdgePropertyBypass}, which 
can be used to clear any visual property.
}
\examples{
\donttest{
unhideEdges()
}
}
\seealso{
{
\link{clearEdgePropertyBypass},
\link{unhideAll}
}
}
