% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{updateApp}
\alias{updateApp}
\title{Update App}
\usage{
updateApp(app, base.url = .defaultBaseUrl)
}
\arguments{
\item{app}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
updateApp()
}
}
