% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{updateStyleDefaults}
\alias{updateStyleDefaults}
\title{Updates the default values of visual properties in a style}
\usage{
updateStyleDefaults(style.name, defaults, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{(char) name for style}

\item{defaults}{(list) a list of visual property default settings}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
server response
}
\description{
Updates visual property defaults, overriding any prior settings. See mapVisualProperty for
the list of visual properties.
}
\examples{
\donttest{
updateStyleDefaults('myStyle',list('node fill color'='#0000FF','node size'=50))
}
}
\seealso{
mapVisualProperty
}
