% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{updateStyleDependencies}
\alias{updateStyleDependencies}
\title{Updates the values of dependencies in a style}
\usage{
updateStyleDependencies(style.name, dependencies, base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{(char) name for style}

\item{dependencies}{A \code{list} of style dependencies, see Available Dependencies
below. Note: each dependency is set by a boolean, TRUE or FALSE (T or F)}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Updates style dependencies, overriding any prior settings.
}
\section{Available Dependencies}{

arrowColorMatchesEdge
nodeCustomGraphicsSizeSync
nodeSizeLocked
}

\examples{
\donttest{
updateStyleDependencies("myStyle",list(nodeSizeLocked=TRUE))
}
}
