% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{addCyEdges}
\alias{addCyEdges}
\title{Add CyEdges}
\usage{
addCyEdges(source.target.list, edgeType = "interacts with",
  directed = FALSE, network = NULL, base.url = .defaultBaseUrl)
}
\description{
Add CyEdges
}
\examples{
\donttest{
addCyEdges(c('sourceNode','targetNode'))
addCyEdges(list(c('s1','t1'),c('s2','t2')))
}
}
