% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{availableApiVersions,missing-method}
\alias{availableApiVersions,missing-method}
\alias{availableApiVersions,character-method}
\title{Available CyREST API Versions}
\usage{
\S4method{availableApiVersions}{missing}()

\S4method{availableApiVersions}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{list} of API versions as \code{character} strings, e.g., "v1"
}
\description{
Get the list of available CyREST API versions
}
\examples{
\donttest{
availableApiVersions()

base.url <- 'http://localhost:1234/v1'
availableApiVersions(base.url)
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
