% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{cloneNetwork,OptionalCyWinClass-method}
\alias{cloneNetwork,OptionalCyWinClass-method}
\title{Clone a Cytoscape Network}
\usage{

  \S4method{cloneNetwork}{OptionalCyWinClass}(obj = CytoscapeWindowFromNetwork(),
  new.title, return.graph = FALSE)
}
\arguments{
\item{obj}{(optional) \code{CytoscapeConnection} or \code{CytoscapeWindow}}

\item{new.title}{Name for the new network}

\item{return.graph}{\code{logical} Whether to include the \code{graphNEL} representation 
of the network as part of the returned \code{CytoscapeWindow}. Warning: this is slow.}
}
\value{
A \code{CytoscapeWindow} object associated with the new network
}
\description{
Makes a copy of a Cytoscape Network with all of its edges and nodes.
}
\examples{
\dontrun{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
copy_of_your_net <- cloneNetwork(cw, "new_copy")
}
}
\seealso{
\code{\link{createNetworkFromSelection}}, 
\code{\link{CytoscapeWindowFromNetwork}}, 
\code{\link{renameNetwork}}
}
\author{
Alexander Pico, Julia Gustavsen
}
