% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{cloneNetwork}
\alias{cloneNetwork}
\title{Clone a Cytoscape Network}
\usage{
cloneNetwork(new.title, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network you want to clone; default is "current" network}

\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
The \code{suid} of the new network
}
\description{
Makes a copy of a Cytoscape Network with all of its edges and nodes.
}
\examples{
\dontrun{
cloneNetwork("cloned network")
}
}
\author{
Alexander Pico, Julia Gustavsen
}
