% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{cloneNetwork}
\alias{cloneNetwork}
\title{Clone a Cytoscape Network}
\usage{
cloneNetwork(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network you want to clone; 
default is "current" network}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
The \code{suid} of the new network
}
\description{
Makes a copy of a Cytoscape Network with all of its edges and nodes.
}
\examples{
\donttest{
cloneNetwork("cloned network")
}
}
\author{
Alexander Pico, Julia Gustavsen
}
