% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.R
\name{closeSession}
\alias{closeSession}
\title{Close Session}
\usage{
closeSession(save.before.closing, filename = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{save.before.closing}{DESCRIPTION}

\item{filename}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
closeSession()
}
}
