% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{commandsHelp}
\alias{commandsHelp}
\title{Commands Help}
\usage{
commandsHelp(cmd.string = "help", base.url = .defaultBaseUrl)
}
\arguments{
\item{cmd.string}{(char) command}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of available commands or args
}
\description{
Using the same syntax as Cytoscape's Command Line Dialog,
this function returns a list of available commands or args.
}
\details{
Works with or without 'help' command prefix. Note that if you ask about a command that doesn't
have any arguments, this function will run the command!
}
\examples{
\donttest{
commandsHelp()
commandsHelp('node')
commandsHelp('node get attribute')
}
}
