% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{copyVisualStyle}
\alias{copyVisualStyle}
\title{Copy Visual Style}
\usage{
copyVisualStyle(from.style, to.style, base.url = .defaultBaseUrl)
}
\arguments{
\item{from.style}{DESCRIPTION}

\item{to.style}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
copyVisualStyle()
}
}
