% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createGraphFromNetwork}
\alias{createGraphFromNetwork}
\title{createGraphFromNetwork}
\usage{
createGraphFromNetwork(obj, title = NA)
}
\value{
A Bioconductor graph object.
}
\description{
Returns the Cytoscape network as a Bioconductor graph.
}
\examples{
\donttest{cw <- CytoscapeWindow('network', graph=makeSimpleGraph())
displayGraph(cw)
layoutNetwork(cw)
g.net1 <- createGraphFromNetwork(cw)

cc <- CytoscapeConnection()
g.net2 <- createGraphFromNetwork(cc, 'network')

g.net3 <- createGraphFromNetwork('network') #default connection

g.net4 <- createGraphFromNetwork() #current network
}
}
\author{
Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
