% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Groups.R
\name{createGroup}
\alias{createGroup}
\title{Create Group}
\usage{
createGroup(group.name, nodes = NULL, nodes.by.col = "name",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{group.name}{DESCRIPTION}

\item{nodes}{List of node names or keyword: selected, unselected or all. Default
is currently selected nodes.}

\item{nodes.by.col}{name of node table column corresponding to provided nodes 
list. Default is 'name'.}

\item{network}{(optional) Name or SUID of the network. Default is the "current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
createGroup()
}
}
