% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createIgraphFromNetwork}
\alias{createIgraphFromNetwork}
\title{Create an igraph network from a Cytoscape network}
\usage{
createIgraphFromNetwork(title = NA, obj = CytoscapeWindowFromNetwork(), ...)
}
\arguments{
\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
(igraph) an igraph network
}
\description{
Takes a Cytoscape network and generates data frames for vertices and edges to
send to the graph_from_data_frame function.
Returns the network.suid and applies the perferred layout set in Cytoscape preferences.
}
\details{
Nodes and edges from the Cytoscape network will be translated into vertices and edges
in igraph. Associated table columns will also be passed to igraph as vertiex and edge attributes.
}
\examples{
\donttest{
createNetworkFromIgraph(g)
}
}
\seealso{
createNetworkFromDataFrames, createIgraphFromNetwork
}
