% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createNetworkFromGraph}
\alias{createNetworkFromGraph}
\title{Create Network From Graph}
\usage{
createNetworkFromGraph(graph, title = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{graph}{DESCRIPTION}

\item{title}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
Creates a Cytoscape network from a Bioconductor graph.
}
\examples{
\donttest{
createNetworkFromGraph(g, "myNetwork")
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
