% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{createNetworkFromGraph}
\alias{createNetworkFromGraph}
\title{createNetworkFromGraph}
\usage{
createNetworkFromGraph(graph, title = NULL, base.url = .defaultBaseUrl)
}
\value{
A Bioconductor graph object.
}
\description{
Creates a Cytoscape network from a Bioconductor graph.
}
\examples{
\donttest{
createNetworkFromGraph(g, "myNetwork")
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
