% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{createWindow}
\alias{createWindow}
\alias{createWindowdeprecated}
\title{DEPRECATED: createWindow}
\usage{
createWindowdeprecated
}
\value{
Network SUID
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
Use the replacement function instead: 
\link{createNetworkFromGraph}
}
