% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestAPI}
\alias{cyrestAPI}
\title{Open Swagger docs for CyREST API}
\usage{
cyrestAPI(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Web page in browser
}
\description{
Opens swagger docs in default browser for a live
instance of CyREST operations.
}
\examples{
\donttest{
cyrestAPI()
}
}
