% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{cyrestPOST}
\alias{cyrestPOST}
\title{CyREST POST}
\usage{
cyrestPOST(operation, parameters = NULL, body = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{operation}{DESCRIPTION}

\item{parameters}{DESCRIPTION}

\item{body}{DESCRIPTION}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
cyrestPOST()
}
}
