% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeApiVersions}
\alias{cytoscapeApiVersions}
\title{Available CyREST API Versions}
\usage{
cytoscapeApiVersions(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A \code{list} of API versions as \code{character} strings, e.g., "v1"
}
\description{
Get the list of available CyREST API versions
}
\examples{
\donttest{
cytoscapeApiVersions()
# [1] "v1"
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
