% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{cytoscapeFreeMemory,missing-method}
\alias{cytoscapeFreeMemory,missing-method}
\alias{cytoscapeFreeMemory,character-method}
\title{Free Up Unused Memory for Cytoscape}
\usage{
\S4method{cytoscapeFreeMemory}{missing}()

\S4method{cytoscapeFreeMemory}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
status message
}
\description{
Manually call Java's garbage collection \code{System.gc()} to free up unused memory. 
This process happens automatically, but may be useful to call explicitly for testing or evaluation purposes.
}
\examples{
\donttest{
cytoscapeFreeMemory()

base.url <- 'http://localhost:1234/v1'
cytoscapeFreeMemory(base.url)
}
}
\author{
Alexander Pico
}
