% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeFreeMemory}
\alias{cytoscapeFreeMemory}
\title{Free Up Unused Memory for Cytoscape}
\usage{
cytoscapeFreeMemory(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
status message
}
\description{
Manually call Java's garbage collection \code{System.gc()} to free up unused memory. 
This process happens automatically, but may be useful to call explicitly for testing or evaluation purposes.
}
\examples{
\donttest{
cytoscapeFreeMemory()
# [1] "Unused memory freed up."
}
}
\author{
Alexander Pico
}
