% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{cytoscapeMemoryStatus,missing-method}
\alias{cytoscapeMemoryStatus,missing-method}
\alias{cytoscapeMemoryStatus,character-method}
\title{Memory Available to Cytoscape}
\usage{
\S4method{cytoscapeMemoryStatus}{missing}()

\S4method{cytoscapeMemoryStatus}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{list} of \code{numeric} values
}
\description{
Returns the memory resources of the server running Cytoscape
}
\examples{
\donttest{
cytoscapeMemoryStatus()

base.url <- 'http://localhost:1234/v1'
cytoscapeMemoryStatus(base.url)
}
}
\author{
Alexander Pico
}
