% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{cytoscapeNumberOfCores,missing-method}
\alias{cytoscapeNumberOfCores,missing-method}
\alias{cytoscapeNumberOfCores,character-method}
\title{Number of Processors Available to Cytoscape}
\usage{
\S4method{cytoscapeNumberOfCores}{missing}()

\S4method{cytoscapeNumberOfCores}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
A \code{numeric} value
}
\description{
Returns the processor resources of the server running Cytoscape
}
\examples{
\donttest{
cytoscapeNumberOfCores()

base.url <- 'http://localhost:1234/v1'
cytoscapeNumberOfCores(base.url)
}
}
\author{
Alexander Pico
}
