% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeNumberOfCores}
\alias{cytoscapeNumberOfCores}
\title{Number of Processors Available to Cytoscape}
\usage{
cytoscapeNumberOfCores(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A \code{numeric} value
}
\description{
Returns the processor resources of the server running Cytoscape
}
\examples{
\donttest{
cytoscapeNumberOfCores()
# [1] 8
}
}
\author{
Alexander Pico
}
