% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapePing}
\alias{cytoscapePing}
\title{Ping Cytoscape}
\usage{
cytoscapePing(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
status message
}
\description{
Tests the connection to Cytoscape via CyREST and verifies that
supported versions of Cytoscape and CyREST API are loaded.
}
\examples{
\donttest{
cytoscapePing()
# [1] "You are connected to Cytoscape!"
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
