% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapePing}
\alias{cytoscapePing}
\title{Ping Cytoscape}
\usage{
cytoscapePing(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional)  URL prefix for CyREST calls}
}
\value{
status message
}
\description{
Test the connection to Cytoscape via CyREST
}
\examples{
\donttest{
cytoscapePing()
# [1] "You are connected to Cytoscape!"
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
