% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{cytoscapeVersionInfo,missing-method}
\alias{cytoscapeVersionInfo,missing-method}
\alias{cytoscapeVersionInfo,character-method}
\title{Cytoscape and CyREST API Versions}
\usage{
\S4method{cytoscapeVersionInfo}{missing}()

\S4method{cytoscapeVersionInfo}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
\code{list} of versions
}
\description{
Returns the versions of the current Cytoscape and CyREST API
}
\examples{
\donttest{
cytoscapeVersionInfo()

base.url <- 'http://localhost:1234/v1'
cytoscapeVersionInfo(base.url)
}
}
\author{
Alexander Pico
}
