% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\name{cytoscapeVersionInfo}
\alias{cytoscapeVersionInfo}
\title{Cytoscape and CyREST API Versions}
\usage{
cytoscapeVersionInfo(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
\code{list} of versions
}
\description{
Returns the versions of the current Cytoscape and CyREST API
}
\examples{
\donttest{
cytoscapeVersionInfo()
 # apiVersion cytoscapeVersion 
 #       "v1" "3.7.0-SNAPSHOT" 
}
}
\author{
Alexander Pico
}
