% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{deleteTableColumn}
\alias{deleteTableColumn}
\title{Delete a table column}
\usage{
deleteTableColumn(column, table = "node", namespace = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{column}{Name of the column to delete}

\item{table}{Name of table, e.g., node (default), edge, network}

\item{namespace}{Namespace of table. Default is "default".}

\item{network}{(optional) Name or SUID of the network. Default is the "current" 
network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A \code{data.frame} of column values
}
\description{
Delete a column from node, edge or network tables.
}
\examples{
\donttest{
deleteTableColumn('node','group')
}
}
