% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{doInitializeSandbox}
\alias{doInitializeSandbox}
\title{doInitializeSandbox}
\usage{
doInitializeSandbox(requester = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{requester}{requester}

\item{base.url}{Ignore unless you need to specify a custom domain, port or 
version to connect to the CyREST API. Default is http://127.0.0.1:1234 and 
the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
doInitializeSandbox
}
\examples{
\donttest{
doInitializeSandbox()
}
}
