% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{doRequestRemote}
\alias{doRequestRemote}
\title{doRequestRemote}
\usage{
doRequestRemote(method, qurl, qbody = NULL, headers = NULL)
}
\arguments{
\item{method}{A string to be converted to the REST query namespace}

\item{qurl}{A named list of values to be converted to REST query parameters}

\item{qbody}{A named list of values to be converted to JSON}

\item{headers}{httr headers}
}
\value{
httr response
}
\description{
Do requests remotely by connecting over Jupyter-Bridge.
}
\examples{
\donttest{
doRequestRemote()
}
}
