% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{doSetSandbox}
\alias{doSetSandbox}
\title{doSetSandbox}
\usage{
doSetSandbox(sandboxToSet, requster = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{sandboxToSet}{sandbox to set}

\item{requster}{requster}

\item{base.url}{base.url (optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\description{
doSetSandbox
}
\examples{
\donttest{
doSetSandbox()
}
}
