% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{enableApp}
\alias{enableApp}
\title{Enable App}
\usage{
enableApp(app, base.url = .defaultBaseUrl)
}
\arguments{
\item{app}{Name of app}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Enable a previously installed and disabled app in Cytoscape.
}
\examples{
\donttest{
enableApp()
}
}
