% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{existing.CytoscapeWindow}
\alias{existing.CytoscapeWindow}
\alias{existing.CytoscapeWindowdeprecated}
\title{DEPRECATED: existing.CytoscapeWindow}
\usage{
existing.CytoscapeWindowdeprecated
}
\value{
Network SUID
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
Use the replacement function instead: 
\link{getNetworkSuid}
}
