% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Commands.R
\name{findRemoteCytoscape}
\alias{findRemoteCytoscape}
\title{findRemoteCytoscape}
\usage{
findRemoteCytoscape()
}
\value{
None
}
\description{
Figure out whether CyREST is local or remote. If remote, we'll 
want to go through Jupyter-Bridge.
}
\examples{
\donttest{
findRemoteCytoscape()
}
}
