% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getAbsSandboxPath}
\alias{getAbsSandboxPath}
\title{getAbsSandboxPath}
\usage{
getAbsSandboxPath(fileLocation)
}
\arguments{
\item{fileLocation}{fileLocation}
}
\value{
file location
}
\description{
getAbsSandboxPath
}
\examples{
\donttest{
getAbsSandboxPath()
}
}
