% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getAppInformation}
\alias{getAppInformation}
\title{Get App Information}
\usage{
getAppInformation(app, base.url = .defaultBaseUrl)
}
\arguments{
\item{app}{Name of app}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
App name, brief description and version.
}
\description{
Retrieve the name, brief description and version of a Cytoscape app.
}
\examples{
\donttest{
getAppInformation()
}
}
