% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getAppUpdates}
\alias{getAppUpdates}
\title{List Apps With Updates}
\usage{
getAppUpdates(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
getAppUpdates()
}
}
