% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getArrowShapes}
\alias{getArrowShapes}
\title{Get Arrow Shapes}
\usage{
getArrowShapes(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A list of \code{character} strings, e.g., 'DIAMOND', 'T', 'ARROW'
}
\description{
Retrieve the names of the currently supported 'arrows' -- the
decorations can (optionally) appear at the ends of edges, adjacent to
the nodes they connect, and conveying information about the nature of
the nodes' relationship.
}
\examples{
\donttest{
getArrowShapes()
}
}
