% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{getDefaultSandbox}
\alias{getDefaultSandbox}
\title{getDefaultSandbox}
\usage{
getDefaultSandbox()
}
\value{
default sandbox
}
\description{
Return whatever is the current default sandbox properties.
}
\examples{
\donttest{
getDefaultSandbox()
}
}
