% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleValues.R
\name{getEdgeColor}
\alias{getEdgeColor}
\title{Get Edge Color}
\usage{
getEdgeColor(edge.names = NULL, network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.names}{List of edge names. Default is NULL for all edges.}

\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Named list of property values
}
\description{
Retrieve the actual line color of specified edges.
}
\examples{
\donttest{
getEdgeColor()
}
}
