% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{getFirstNeighbors,missing-method}
\alias{getFirstNeighbors,missing-method}
\alias{getFirstNeighbors,CytoscapeConnectionClass-method}
\alias{getFirstNeighbors,CytoscapeWindowClass-method}
\title{Get list of nodes neighboring provided list}
\usage{
\S4method{getFirstNeighbors}{missing}(node.names, as.nested.list = FALSE)

\S4method{getFirstNeighbors}{CytoscapeConnectionClass}(obj, node.names,
  as.nested.list = FALSE)

\S4method{getFirstNeighbors}{CytoscapeWindowClass}(obj, node.names,
  as.nested.list = FALSE)
}
\arguments{
\item{node.names}{A \code{list} of node names from the \code{name} column of the \code{node table}}

\item{as.nested.list}{\code{logical} Whether to return lists of neighbors per query node}

\item{obj}{(optional) \code{CytoscapeConnection} or \code{CytoscapeWindow}}
}
\value{
A list of unique node names, optionally nested per query node name.
}
\description{
Returns a non-redundan list of first
neighbors of the supplied list of nodes.
}
\examples{
/dontrun {
# first, delete existing windows to save memory:
deleteAllWindows(CytoscapeConnection())

cw <- CytoscapeWindow ('getFirstNeighbors.test', graph=makeSimpleGraph())
displayGraph (cw)
redraw (cw)
layoutNetwork(cw, 'grid')
print (getFirstNeighbors (cw, 'A'))
selectNodes (cw, getFirstNeighbors (cw, 'A'))  # note that A is not selected
}
}
\seealso{
selectNodes  
selectFirstNeighbors
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
