% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{methods}
\name{getGraphFromNetwork,missing-method}
\alias{getGraphFromNetwork,missing-method}
\alias{getGraphFromNetwork,CytoscapeConnectionClass-method}
\alias{getGraphFromNetwork,CytoscapeWindowClass-method}
\title{getGraphFromNetwork}
\usage{
\S4method{getGraphFromNetwork}{missing}(title = NA)

\S4method{getGraphFromNetwork}{CytoscapeConnectionClass}(obj, title = NA)

\S4method{getGraphFromNetwork}{CytoscapeWindowClass}(obj)
}
\value{
A Bioconductor graph object.
}
\description{
Returns the Cytoscape network as a Bioconductor graph.
}
\examples{
\dontrun{
g.net <- getGraphFromNetwork() #current network

g.net1 <- getGraphFromNetwork('network1')

cc <- CytoscapeConnection()
g.net2 <- getGraphFromNetwork(cc, 'network2')

cw <- CytoscapeWindow('network3', graph=makeSimpleGraph())
displayGraph(cw)
layoutNetwork(cw)
g.net3 <- getGraphFromNetwork(cw)

}
}
\author{
Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
