% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{getInstalledApps}
\alias{getInstalledApps}
\title{List Installed Apps}
\usage{
getInstalledApps(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A list of installed app names, versions and statuses
}
\description{
Retrieve list of currently installed apps in Cytoscape.
}
\examples{
\donttest{
getInstalledApps()
}
}
