% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getLineStyles}
\alias{getLineStyles}
\title{Get Line Styles}
\usage{
getLineStyles(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
A list of \code{character} strings, e.g., 'SOLID', 'DOT'
}
\description{
Retrieve the names of the currently supported line types
-- values which can be used to render edges, and thus
can be used in calls to 'setEdgeLineStyleRule'.
}
\examples{
\donttest{
getLineStyles()
}
}
