% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetworkList}
\alias{getNetworkList}
\title{Get the list of Cytoscape networks}
\usage{
getNetworkList(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
\code{list}
}
\description{
Returns the list of Cytoscape network names in the current Cytoscape session
}
\examples{
\donttest{
getNetworkList()
# 3
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
