% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetworkName}
\alias{getNetworkName}
\title{Get the name of a network}
\usage{
getNetworkName(suid = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{suid}{SUID of the network; default is current network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
network name
}
\description{
Get the name of a network
}
\examples{
\donttest{
getNetworkName()
getNetworkName(1111)
}
}
