% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{getNetworkName}
\alias{getNetworkName}
\title{Get the name of a network}
\usage{
getNetworkName(suid = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{suid}{(optional) SUID of the network; default is current network. If a
name is provided, then it is validated and returned.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
network name
}
\description{
Retrieve the title of a network
}
\examples{
\donttest{
getNetworkName()
getNetworkName(1111)
}
}
