% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getNodePosition}
\alias{getNodePosition}
\title{DEFUNCT: getNodePosition}
\usage{
Use the replacement function instead:
\code{layoutCopycat}
}
\description{
This function is defunct and will be removed in the next release.
}
