% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{getNodePosition}
\alias{getNodePosition}
\alias{getNodePositiondefunct}
\title{DEFUNCT: getNodePosition}
\usage{
getNodePositiondefunct
}
\value{
None
}
\description{
This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{layoutCopycat}
}
