% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getNodeShapes}
\alias{getNodeShapes}
\title{Get Node Shapes}
\usage{
getNodeShapes(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{DESCRIPTION}
}
\value{
A list of \code{character} strings, e.g. 'ELLIPSE', 'RECTANGLE'
}
\description{
Retrieve the names of the currently supported node shapes,
which can then be used in calls to setNodeShapeRule and
setDefaultVizMapValue.
}
\examples{
\donttest{
getNodeShapes()
}
}
