% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{getStyleDependencies}
\alias{getStyleDependencies}
\title{Get the values of dependencies in a style}
\usage{
getStyleDependencies(style.name = "default", base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
server response
}
\description{
Updates style dependencies, overriding any prior settings.
}
\section{Available Dependencies}{

arrowColorMatchesEdge
nodeCustomGraphicsSizeSync
nodeSizeLocked
}

\examples{
\donttest{
getStyleDependencies("myStyle")
}
}
