% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{getTableColumnNames}
\alias{getTableColumnNames}
\title{Get Table Column Names}
\usage{
getTableColumnNames(table = "node", namespace = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{name of table, e.g., node, edge, network; default is "node"}

\item{namespace}{namespace of table, e.g., default}

\item{network}{(optional) Name or SUID of the network. Default is the "current" 
network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
list of column names
}
\description{
Retrieve the names of all columns in a table
}
\examples{
\donttest{
getTableColumnNames()
getTableColumnNames('edge')
getTableColumnNames('network')
}
}
