% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{getTableColumns}
\alias{getTableColumns}
\title{Get table column values}
\usage{
getTableColumns(table, columns = NULL, namespace = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table}{name of table, e.g., node, edge, network}

\item{columns}{names of columns to retrieve values from as list object or comma-separated list; default is all columns}

\item{namespace}{namespace of table; default is "default"}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
A \code{data.frame} of column values
}
\description{
Get table column values
}
\details{
The 'suid' column is always retrieved along with specified columns. 
The 'suid' values are used as \code{row.names} in the returned \code{data.frame}.
}
\examples{
\donttest{
example(createNetworkFromDataFrames)

getTableColumns('node','group')
}
}
