% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Styles.R
\name{getVisualPropertyNames}
\alias{getVisualPropertyNames}
\title{Get Visual Property Names}
\usage{
getVisualPropertyNames(base.url = .defaultBaseUrl)
}
\arguments{
\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
List of names
}
\description{
Retrieve the names of all possible visual properties.
}
\examples{
\donttest{
getVisualPropertyNames()
}
}
