% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables.R
\name{loadTableData}
\alias{loadTableData}
\title{Loads data into Cytoscape tables keyed by row}
\usage{
loadTableData(data, data.key.column = "row.names", table = "node",
  table.key.column = "name", network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{data}{(data.frame) each row is a node and columns contain node attributes}

\item{data.key.column}{(char) name of data.frame column to use as key; default is "row.names"}

\item{table}{(char) name of Cytoscape table to load data into, e.g., node, edge or network; default is "node"}

\item{table.key.column}{(char) name of Cytoscape table column to use as key; default is "name"}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
This function loads data into Cytoscape node/edge/network tables provided a
common key, e.g., name. Data.frame column names will be used to set Cytoscape table column
names.
}
\details{
Numeric (or integer) values will be stored as Doubles in Cytoscape tables.
Character or mixed values will be stored as Strings. Logical values will be stored as Boolean.
Existing columns with the same names will be overwritten.
}
