% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Apps.R
\name{openAppStore}
\alias{openAppStore}
\title{Open App Store Page}
\usage{
openAppStore(app, base.url = .defaultBaseUrl)
}
\arguments{
\item{app}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
openAppStore()
}
}
