% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.R
\name{openSession}
\alias{openSession}
\title{Open Session File or URL}
\usage{
openSession(file.location, base.url = .defaultBaseUrl)
}
\arguments{
\item{file.location}{File path or URL (with 'http' prefix)}
}
\description{
Open Session File or URL
}
