% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.R
\name{openSession}
\alias{openSession}
\title{Open Session File or URL}
\usage{
openSession(file.location = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{file.location}{File path or URL (with 'http' or 'https' prefix). Default 
is a sample session file.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Open a session file or URL. This will clear all networks, tables
and styles associated with current session. Be sure to \link{saveSession} first.
}
\examples{
\donttest{
openSession('/fullpath/mySession.CYS')
}
}
\seealso{
saveSession
}
