% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{ping,missing-method}
\alias{ping,missing-method}
\alias{ping,character-method}
\title{Ping Cytoscape}
\usage{
\S4method{ping}{missing}()

\S4method{ping}{character}(base.url)
}
\arguments{
\item{base.url}{(optional) URL prefix for CyREST calls}
}
\value{
status message
}
\description{
Test the connection to Cytoscape via CyREST
}
\examples{
\donttest{
ping()

base.url <- 'http://localhost:1234/v1'
ping(base.url)
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
