% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoscapeSystem.R
\docType{methods}
\name{ping,missing-method}
\alias{ping,missing-method}
\alias{ping,CytoscapeConnectionClass-method}
\title{Ping Cytoscape}
\usage{
\S4method{ping}{missing}()

\S4method{ping}{CytoscapeConnectionClass}(obj)
}
\arguments{
\item{obj}{(optional) \code{CytoscapeConnection}}
}
\value{
status message
}
\description{
Test the connection to Cytoscape via CyREST
}
\examples{
\donttest{
ping()

cy <- CytoscapeConnection()
ping(cy)
}
}
\author{
Alexander Pico, Tanja Muetze, Georgi Kolishovski, Paul Shannon
}
