% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3.R
\docType{package}
\name{RCy3}
\alias{RCy3}
\alias{RCy3-package}
\title{RCy3: Functions to Access and Control Cytoscape}
\description{
Vizualize, analyze and explore networks using Cytoscape via R.
}
\details{
To learn more about RCy3, start with the vignettes:
\code{browseVignettes("RCy3")}
}
