% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-package.r
\docType{package}
\name{RCy3}
\alias{RCy3}
\alias{RCy3-package}
\title{RCy3: Display and manipulate graphs in Cytoscape >= 3.3.0}
\description{
The RCy3 package provides functions to run Cytoscape from R.
}
\details{
#' To learn more about RCy3, start with the vignettes:
\code{browseVignettes(package = "RCy3")}
}
