% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename a network}
\usage{
renameNetwork(title, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{title}{New name for the network}

\item{network}{name or suid of the network that you want to rename; default is "current" network}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Sets a new name for this network
}
\details{
Duplicate network names are not allowed
}
\examples{
\dontrun{
renameNetwork("renamed network")
}
}
\author{
Alexander Pico, Julia Gustavsen
}
