% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{renameNetwork}
\alias{renameNetwork}
\title{Rename a network}
\usage{
renameNetwork(new.title, network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{new.title}{New name for the network}

\item{network}{name or suid of the network that you want set as current}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
None
}
\description{
Sets a new name for this network
}
\examples{
\dontrun{
renameNetwork("renamed network")
}
}
\author{
Alexander Pico, Julia Gustavsen
}
