% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{restoreLayout}
\alias{restoreLayout}
\alias{restoreLayoutdefunct}
\alias{getGraph}
\alias{restoreLayoutdefunct}
\title{DEFUNCT: restoreLayout}
\usage{
restoreLayoutdefunct

restoreLayoutdefunct
}
\value{
None

None
}
\description{
This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{layoutCopycat}

This function is defunct and will be removed in the next release.
Use the replacement function instead: 
\link{createGraphFromNetwork}
}
