% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-notebook.R
\name{runningRemoteCheck}
\alias{runningRemoteCheck}
\title{runningRmoteCheck}
\usage{
runningRemoteCheck(newState = NULL)
}
\arguments{
\item{newState}{new state of running remote}
}
\value{
old state
}
\description{
runningRmoteCheck
}
\examples{
\donttest{
runningRmoteCheck()
}
}
