% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxGetFileInfo}
\alias{sandboxGetFileInfo}
\title{sandboxGetFileInfo}
\usage{
sandboxGetFileInfo(fileName, sandboxName = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{fileName}{Name of file whose metadata to return ... can be sandbox-relative path ... ``.`` returns metadata on sandbox itself}

\item{sandboxName}{Name of sandbox containing file. None means "the current sandbox".}

\item{base.url}{Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://127.0.0.1:1234 and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
dict: {'filePath': <full path on Cytoscape workstation>, 'modifiedTime': <last changed time, '' if file doesn't exist>, 'isFile': <True if file, False if directory>}
}
\description{
sandboxGetFileInfo
}
\examples{
\donttest{
sandboxGetFileInfo()
}
}
