% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{sandboxInitializer}
\alias{sandboxInitializer}
\title{sandboxInitializer}
\usage{
sandboxInitializer(newSandbox = NULL, ...)
}
\arguments{
\item{newSandbox}{newSandbox}

\item{...}{...}
}
\value{
sandbox
}
\description{
Start with a sandbox template and update properties using whatever is found in the new_sandbox.
}
\examples{
\donttest{
sandboxInitializer()
}
}
