% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxOp}
\alias{sandboxOp}
\title{sandboxOp}
\usage{
sandboxOp(command, sandboxName, fileName = NULL, base.url = .defaultBaseUrl)
}
\value{
sandboxOp
}
\description{
sandboxOp
}
\examples{
\donttest{
sandboxOp()
}
}
