% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxRemove}
\alias{sandboxRemove}
\title{sandboxRemove}
\usage{
sandboxRemove(sandboxName = NULL, base.url = .defaultBaseUrl)
}
\value{
sandboxRemove
}
\description{
sandboxRemove
}
\examples{
\donttest{
sandboxRemove()
}
}
