% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxRemoveFile}
\alias{sandboxRemoveFile}
\title{sandboxRemoveFile}
\usage{
sandboxRemoveFile(fileName, sandboxName = NULL, base.url = .defaultBaseUrl)
}
\value{
sandboxRemoveFile
}
\description{
sandboxRemoveFile
}
\examples{
\donttest{
sandboxRemoveFile()
}
}
