% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxRemoveFile}
\alias{sandboxRemoveFile}
\title{sandboxRemoveFile}
\usage{
sandboxRemoveFile(fileName, sandboxName = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{fileName}{Name of file to delete (as absolute path or sandbox-relative path)}

\item{sandboxName}{Name of sandbox containing file. None means "the current sandbox".}

\item{base.url}{Ignore unless you need to specify a custom domain, port or version to connect to the CyREST API. Default is http://127.0.0.1:1234 and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
dict: {'filePath': <file's absolute path in Cytoscape workstation>, 'existed': True if file existed before being deleted}
}
\description{
sandboxRemoveFile
}
\examples{
\donttest{
sandboxRemoveFile()
}
}
