% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sandbox.R
\name{sandboxSet}
\alias{sandboxSet}
\title{sandboxSet}
\usage{
sandboxSet(
  sandboxName,
  copySamples = TRUE,
  reinitialize = TRUE,
  base.url = .defaultBaseUrl
)
}
\value{
sandboxPath
}
\description{
sandboxSet
}
\examples{
\donttest{
sandboxSet()
}
}
