% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Session.R
\name{saveSession}
\alias{saveSession}
\title{Saves the current Cytoscape session}
\usage{
saveSession(filename, obj = CytoscapeConnection())
}
\arguments{
\item{filename}{(char) name of the session file to save}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Saves the current Cytoscape session
}
\details{
Saves session as a CYS file.
}
\examples{
\donttest{
saveSession('myFirstSession')
}
}
