% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSelection.R
\name{selectAllEdges}
\alias{selectAllEdges}
\title{Select all edges}
\usage{
selectAllEdges(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Selects all edges in a specified network.
}
\description{
Selects all edges in a Cytoscape Network
}
\examples{
\donttest{
cw <- CytoscapeWindow('new.demo', new('graphNEL'))
selectAllEdges(cw)
}
}
\author{
Alexander Pico, Julia Gustavsen
}
