% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkSelection.R
\name{selectEdgesConnectingSelectedNodes}
\alias{selectEdgesConnectingSelectedNodes}
\title{Select the edges connecting selected nodes in Cytoscape Network}
\usage{
selectEdgesConnectingSelectedNodes(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
Lists of SUIDs for selected nodes and edges
}
\description{
Selects edges in a Cytoscape Network connecting the selected 
nodes, including self loops connecting single nodes.
}
\examples{
\dontrun{
selectEdgesConnectingSelectedNodes()
}
}
\author{
Alexander Pico, Julia Gustavsen
}
