% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\docType{methods}
\name{selectFirstNeighbors,missing-method}
\alias{selectFirstNeighbors,missing-method}
\alias{selectFirstNeighbors,CytoscapeConnectionClass-method}
\alias{selectFirstNeighbors,CytoscapeWindowClass-method}
\title{Select first neighbor nodes}
\usage{
\S4method{selectFirstNeighbors}{missing}(direction = "any")

\S4method{selectFirstNeighbors}{CytoscapeConnectionClass}(obj,
  direction = "any")

\S4method{selectFirstNeighbors}{CytoscapeWindowClass}(obj, direction = "any")
}
\arguments{
\item{direction}{direction of connections to neighbors to follow, e.g., incoming, outgoing, undirected, or any (default)}

\item{network}{name or suid of the network; default is "current" network}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
list of names of selected nodes, including original selection
}
\description{
Select nodes directly connected to currently selected nodes. Can
specify connection directionality using the direction param.
}
\examples{
\donttest{
selectFirstNeighbors()
selectFirstNeighors('outgoing')
selectFirstNeighors('incoming')
}
}
