% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-deprecated.R
\name{sendNodes}
\alias{sendNodes}
\alias{sendNodesdefunct}
\title{DEFUNCT: sendNodes}
\usage{
sendNodesdefunct
}
\value{
None
}
\description{
This function is only provided for compatibility with older
versions of RCy3 and will be defunct and removed in the next releases.
}
