% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Networks.R
\name{setCurrentNetwork}
\alias{setCurrentNetwork}
\title{Set current network}
\usage{
setCurrentNetwork(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{name or suid of the network that you want set as current}

\item{base.url}{cyrest base url for communicating with cytoscape}
}
\value{
server response
}
\description{
Selects the given network as "current"
}
\examples{
\donttest{
setCurrentNetwork('MyNetwork')
}
}
