% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{setCurrentSandbox}
\alias{setCurrentSandbox}
\title{setCurrentSandbox}
\usage{
setCurrentSandbox(sandboxName, sandboxPath)
}
\arguments{
\item{sandboxName}{sandboxName}

\item{sandboxPath}{sandboxPath}
}
\value{
current sandbox
}
\description{
Set and return the current sandbox name and path.
}
\examples{
\donttest{
setCurrentSandbox()
}
}
