% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworkViews.R
\name{setCurrentView}
\alias{setCurrentView}
\title{Set Current View}
\usage{
setCurrentView(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of a network or view. The first 
(presummably only) view associated a network is used if a specific view 
SUID is not provided.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set which network view is "current".
}
\details{
Takes first (presumably only) view associated with provided network
}
\examples{
\donttest{
setCurrentView()
}
}
