% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{setDefaultSandbox}
\alias{setDefaultSandbox}
\title{setDefaultSandbox}
\usage{
setDefaultSandbox(newSandbox = NULL, ...)
}
\arguments{
\item{newSandbox}{newSandbox}

\item{...}{...}
}
\value{
default sandbox
}
\description{
Set and return the sandbox properties to be used as a default, probably based on whether running remote.
}
\examples{
\donttest{
setDefaultSandbox()
}
}
