% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{setDefaultSandboxPath}
\alias{setDefaultSandboxPath}
\title{setDefaultSandboxPath}
\usage{
setDefaultSandboxPath(newPath)
}
\value{
default sandbox path
}
\description{
setDefaultSandboxPath
}
\examples{
\donttest{
setDefaultSandboxPath()
}
}
