% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeLineStyleMapping}
\alias{setEdgeLineStyleMapping}
\title{Set Edge Line Style Mapping}
\usage{
setEdgeLineStyleMapping(edge.attribute.name, attribute.values, line.styles,
  default.style = "SOLID", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.attribute.name}{DESCRIPTION}

\item{attribute.values}{DESCRIPTION}

\item{line.styles}{DESCRIPTION}

\item{default.style}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeLineStyleMapping()
}
}
