% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeLineWidthMapping}
\alias{setEdgeLineWidthMapping}
\title{Set Edge Line Width Mapping}
\usage{
setEdgeLineWidthMapping(edge.attribute.name, attribute.values, line.widths,
  mode = "interpolate", default.width = 1, style.name = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{edge.attribute.name}{DESCRIPTION}

\item{attribute.values}{DESCRIPTION}

\item{line.widths}{DESCRIPTION}

\item{mode}{DESCRIPTION}

\item{default.width}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeLineWidthMapping()
}
}
