% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeSourceArrowMapping}
\alias{setEdgeSourceArrowMapping}
\title{Set Edge Source Arrow Mapping}
\usage{
setEdgeSourceArrowMapping(table.column, table.column.values, shapes,
  default.shape = "ARROW", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{shapes}{DESCRIPTION}

\item{default.shape}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeSourceArrowMapping()
}
}
