% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setEdgeTargetArrowMapping}
\alias{setEdgeTargetArrowMapping}
\title{Set Edge Target Arrow Mapping}
\usage{
setEdgeTargetArrowMapping(table.column, table.column.values, arrows,
  default.arrow = "ARROW", style.name = "default", network = NULL,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{arrows}{DESCRIPTION}

\item{default.arrow}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setEdgeTargetArrowMapping()
}
}
