% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDefaults.R
\name{setNodeLabelPositionDefault}
\alias{setNodeLabelPositionDefault}
\title{Set Node Label Position Default}
\usage{
setNodeLabelPositionDefault(
  new.nodeAnchor,
  new.graphicAnchor,
  new.justification,
  new.xOffset,
  new.yOffset,
  style.name = NULL,
  base.url = .defaultBaseUrl
)
}
\arguments{
\item{style.name}{Name of style; default is "default" style.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}

\item{new.position}{value for position}
}
\value{
None
}
\description{
Set the default node label position
}
\examples{
\donttest{
setNodeLabelPositionDefaultt("S","C","c",0.00,0.00)
}
}
