% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleMappings.R
\name{setNodeSizeMapping}
\alias{setNodeSizeMapping}
\title{Set Node Size Mapping}
\usage{
setNodeSizeMapping(table.column, table.column.values = NULL, sizes = NULL,
  mapping.type = "c", default.size = NULL, style.name = "default",
  network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{table.column}{DESCRIPTION}

\item{table.column.values}{DESCRIPTION}

\item{sizes}{DESCRIPTION}

\item{mapping.type}{(char) continuous, discrete or passthrough (c,d,p); default is continuous}

\item{default.size}{DESCRIPTION}

\item{style.name}{DESCRIPTION}

\item{network}{DESCRIPTION}

\item{base.url}{DESCRIPTION}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
\donttest{
setNodeSizeMapping()
}
}
