% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCy3-sandbox.R
\name{setSandboxReinitialize}
\alias{setSandboxReinitialize}
\title{setSandboxReinitialize}
\usage{
setSandboxReinitialize(doReinitialize = TRUE)
}
\value{
sandbox reinitialize
}
\description{
setSandboxReinitialize
}
\examples{
\donttest{
setCurrentSandbox()
}
}
