% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{setStyleDependencies}
\alias{setStyleDependencies}
\title{Set Style Dependencies}
\usage{
setStyleDependencies(style.name = "default", dependencies,
  base.url = .defaultBaseUrl)
}
\arguments{
\item{style.name}{Name of style; default is "default" style}

\item{dependencies}{A \code{list} of style dependencies, see Available Dependencies
below. Note: each dependency is set by a boolean, TRUE or FALSE (T or F)}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
server response
}
\description{
Sets the values of dependencies in a style, overriding any prior settings.
}
\section{Available Dependencies}{

arrowColorMatchesEdge
nodeCustomGraphicsSizeSync
nodeSizeLocked
}

\examples{
\donttest{
setStyleDependencies("myStyle",list(nodeSizeLocked=TRUE))
}
}
