% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserInterface.R
\name{showGraphicsDetails}
\alias{showGraphicsDetails}
\title{Show Graphics Details}
\usage{
showGraphicsDetails(new.value, base.url = .defaultBaseUrl)
}
\arguments{
\item{new.value}{\code{boolean} Whether to show graphic details}

\item{base.url}{DESCRIPTION}
}
\value{
None
}
\description{
Regardless of the current zoom level and network size,
display (or hide) graphics details, e.g., node labels.
}
\details{
Displaying graphics details on a very large network will affect pan
and zoom performance, depending on your available RAM. 
See \link{cytoscapeMemoryStatus}.
}
\examples{
\donttest{
showGraphicsDetails(TRUE)
}
}
