% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleDependencies.R
\name{syncNodeCustomGraphicsSize}
\alias{syncNodeCustomGraphicsSize}
\title{Sync Node Custom Graphics Size}
\usage{
syncNodeCustomGraphicsSize(new.state, style.name = "default",
  base.url = .defaultBaseUrl)
}
\arguments{
\item{new.state}{(Boolean) Whether to sync node custom graphics size}

\item{style.name}{Name of style; default is "default" style}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Set a boolean value to have the size of custom graphics match that
of the node.
}
\examples{
\donttest{
syncNodeCustomGraphicsSize(TRUE)
}
}
