% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StyleBypasses.R
\name{unhideAll}
\alias{unhideAll}
\title{Unhide All}
\usage{
unhideAll(network = NULL, base.url = .defaultBaseUrl)
}
\arguments{
\item{network}{(optional) Name or SUID of the network. Default is the 
"current" network active in Cytoscape.}

\item{base.url}{(optional) Ignore unless you need to specify a custom domain,
port or version to connect to the CyREST API. Default is http://localhost:1234
and the latest version of the CyREST API supported by this version of RCy3.}
}
\value{
None
}
\description{
Unhide all previously hidden nodes and edges, by 
clearing the Visible property bypass value.
}
\details{
This method sets node and edge visibility bypass to true, overriding
any defaults or mappings. Pending CyREST updates, this method will
ultimately call the generic function, \link{clearEdgePropertyBypass}, which 
can be used to clear any visual property.
}
\examples{
\donttest{
unhideAll()
}
}
\seealso{
{
\link{clearEdgePropertyBypass},
\link{unhideNodes}
\link{unhideEdges}
}
}
